+++ docs/newsfragments/+core.fixed.md
@@ -1,1 +1,0 @@
-Refactored internal import helper to use , fixing support for nested module imports (e.g., ).
