+++ docs/newsfragments/+geom.added.md
@@ -1,1 +1,0 @@
-Added a new fragment detection tool () supporting both Geometric (covalent radii) and Bond Order (GFN2-xTB) methodologies.
