+++ docs/newsfragments/+5ce7a91b.cli.added.md
@@ -1,1 +1,0 @@
-feat(cli): drop dependency on click
