+++ docs/newsfragments/+0a91c98c.plt_neb.added.md
@@ -1,1 +1,0 @@
-feat(pltneb): Add 'ira-kmax' to tweak settings for IRA
