### 2025-12-23 3f95986
feat(eonmlf): start by logging ini details
