### 2025-12-23 9f3006d
feat(eonmlflow): hydrate and log with sections
