+++ rgpycrumbs/eon/_mlflow/log_params.py
@@ -11,35 +11,33 @@ eon_config = _import_from_parent_env("eon.config")
 
 def log_config_ini(conf_ini: Path = Path("config.ini"), *, w_artifact: bool = True):
     """
-    Logs the full EON configuration state, including schema defaults.
+    Logs the hydrated EON configuration and highlights user-provided overrides.
 
-    This function reconstructs the hierarchical configuration by combining
-    the EON schema with user-defined overrides. It preserves the 'Section/Key'
-    provenance required for rigorous simulation tracking.
+    This function performs a three-way merge between the EON schema, the
+    hydrated defaults, and the user-provided config.ini. It ensures that
+    scientific provenance remains intact while providing a focused view of
+    modified hyperparameters.
     """
-    # 1. Instantiate the class to load the config.yaml schema
     econf = eon_config.ConfigClass()
 
-    # 2. Build a local parser to hold the hydrated state
+    # Build a local parser to hold the hydrated state
     # This emulates the logic inside econf.init but preserves the parser object
     hydrated_parser = configparser.ConfigParser()
+    user_parser = configparser.ConfigParser()
 
-    # 3. Apply defaults from the EON schema (provenance of defaults)
+    # Populate hydrated state with defaults from schema
     for section in econf.format:
-        section_name = section.name
-        if not hydrated_parser.has_section(section_name):
-            hydrated_parser.add_section(section_name)
-
+        if not hydrated_parser.has_section(section.name):
+            hydrated_parser.add_section(section.name)
         for config_key in section.keys:
-            # We use the default value defined in the EON source
-            hydrated_parser.set(section_name, config_key.name, str(config_key.default))
+            hydrated_parser.set(section.name, config_key.name, str(config_key.default))
 
-    # 4. Apply user overrides from the config.ini file
+    # Read user overrides if they exist
     if conf_ini.exists():
+        user_parser.read(str(conf_ini.absolute()))
         hydrated_parser.read(str(conf_ini.absolute()))
 
-    # 5. Log the combined state to MLflow
-    # We map 'kind' strings to ConfigParser getters for type-safe logging
+    # Map 'kind' strings to ConfigParser getters for type-safe logging
     type_getters = {
         "int": hydrated_parser.getint,
         "float": hydrated_parser.getfloat,
@@ -47,22 +45,33 @@ def log_config_ini(conf_ini: Path = Path("config.ini"), *, w_artifact: bool = Tr
         "string": hydrated_parser.get,
     }
 
+    # Log all parameters and focus on overrides
     for section in econf.format:
         section_name = section.name
         for config_key in section.keys:
             key_name = config_key.name
+            full_key = f"{section_name}/{key_name}"
             getter = type_getters.get(config_key.kind, hydrated_parser.get)
 
             try:
-                # Retrieve the hydrated value (default or user-override)
+                # Always log the hydrated value for the full record
                 val = getter(section_name, key_name)
-                mlflow.log_param(f"{section_name}/{key_name}", val)
+                mlflow.log_param(full_key, val)
+
+                # Focused logging: If the user explicitly provided this in the INI
+                if user_parser.has_option(section_name, key_name):
+                    mlflow.log_param(f"Overrides/{full_key}", val)
+
             except (ValueError, configparser.Error):
-                # Fallback to raw string logging upon type mismatch
-                mlflow.log_param(
-                    f"{section_name}/{key_name}",
-                    hydrated_parser.get(section_name, key_name),
-                )
+                raw_val = hydrated_parser.get(section_name, key_name)
+                mlflow.log_param(full_key, raw_val)
+                if user_parser.has_option(section_name, key_name):
[…]
