### 2025-12-23 8a8ac86
chore(consplit): rework with alignment options
