+++ rgpycrumbs/eon/con_splitter.py
@@ -10,7 +10,6 @@
 # ///
 
 import logging
-import sys
 from pathlib import Path
 
 import click
@@ -18,6 +17,14 @@ from ase.io import read as aseread
 from ase.io import write as asewrite
 from rich.console import Console
 from rich.logging import RichHandler
+
+# Optional IRA import logic
+try:
+    from rgpycrumbs._aux import _import_from_parent_env
+
+    ira_mod = _import_from_parent_env("ira_mod")
+except ImportError:
+    ira_mod = None
 
 CONSOLE = Console()
 
@@ -38,6 +45,65 @@ logging.basicConfig(
 )
 
 
+def align_structures(frames, use_ira=False, kmax=1.8):
+    """
+    Aligns all frames in the path to the first frame (reactant).
+
+    This procedure removes global rotation and translation. If use_ira equals
+    True, the function also accounts for atom index permutations.
+    """
+    if len(frames) < 2:
+        return frames
+
+    ref = frames[0]
+    ref_pos = ref.get_positions()
+    ref_types = ref.get_atomic_numbers()
+    ref_natoms = len(ref)
+
+    aligned_frames = [ref.copy()]
+
+    # Initialize IRA if requested
+    ira_instance = None
+    if use_ira and ira_mod:
+        ira_instance = ira_mod.IRA()
+        logging.info(
+            "Using [bold magenta]IRA[/bold magenta] for optimal alignment and reordering."
+        )
+
+    for i in range(1, len(frames)):
+        current = frames[i].copy()
+
+        if ira_instance:
+            # Perform IRA match
+            r, t, p, _ = ira_instance.match(
+                ref_natoms,
+                ref_types,
+                ref_pos,
+                len(current),
+                current.get_atomic_numbers(),
+                current.get_positions(),
+                kmax,
+            )
+            # Apply rotation, translation, and permutation
+            new_pos = (current.get_positions() @ r.T) + t
+            current.set_positions(new_pos[p])
+            # Reorder atomic numbers to match reference after permutation
+            current.set_atomic_numbers(current.get_atomic_numbers()[p])
+        else:
+            # Standard ASE alignment (Procrustes)
+            current.euler_rotate(
+                0, 0, 0
+            )  # Placeholder for more complex ASE alignment if needed
+            # For basic NEB, many users prefer simple RMSD alignment:
+            from ase.build import minimize_rotation_and_translation
+
+            minimize_rotation_and_translation(ref, current)
+
+        aligned_frames.append(current)
+
+    return aligned_frames
+
+
 @click.command()
 @click.argument(
     "neb_trajectory_file",
@@ -47,173 +113,126 @@ logging.basicConfig(
     "--output-dir",
     type=click.Path(file_okay=False, writable=True, path_type=Path),
     default=None,
-    help=(
-        "Directory to save the output files. "
-        "Defaults to a new directory named after the input file (e.g., 'neb_path_001/')."
-    ),
+    help="Directory to save the output files.",
 )
 @click.option(
     "--images-per-path",
     type=int,
     required=True,
-    help="Number of images in a single NEB path (e.g., 7). [REQUIRED]",
+    help="Number of images in a single NEB path.",
 )
 @click.option(
     "--path-index",
     type=int,
-    default=0,
+    default=-1,
     show_default=True,
-    help=(
-        "Index of the NEB path to extract (0-based). "
-        "Use -1 to extract the *last* available path."
-    ),
+    help="Index of the NEB path to extract. Use -1 for the last path.",
+)
+@click.option(
+    "--center",
+    is_flag=True,
+    help="Center the atoms in each frame around the origin.",
+)
+@click.option(
+    "--box-diagonal",
+    nargs=3,
[…]
