### 2025-12-23 19e8b07
feat(consplit): rework to add more alignment
