+++ rgpycrumbs/eon/con_splitter.py
@@ -10,9 +10,12 @@
 # ///
 
 import logging
+import sys
+from enum import Enum
 from pathlib import Path
 
 import click
+from ase.build import minimize_rotation_and_translation
 from ase.io import read as aseread
 from ase.io import write as asewrite
 from rich.console import Console
@@ -45,63 +48,59 @@ logging.basicConfig(
 )
 
 
-def align_structures(frames, use_ira=False, kmax=1.8):
-    """
-    Aligns all frames in the path to the first frame (reactant).
+class AlignMode(Enum):
+    """Defines structural alignment strategies."""
 
-    This procedure removes global rotation and translation. If use_ira equals
-    True, the function also accounts for atom index permutations.
-    """
-    if len(frames) < 2:
+    NONE = "none"
+    ALL = "all"  # Align every frame to the reactant
+    ENDPOINTS = "endpoints"  # Align only reactant and product to each other
+
+
+def perform_alignment(ref, target, use_ira=False, kmax=1.8):
+    """Performs the actual rotation/translation match between two structures."""
+    if use_ira and ira_mod:
+        ira_instance = ira_mod.IRA()
+        r, t, p, _ = ira_instance.match(
+            len(ref),
+            ref.get_atomic_numbers(),
+            ref.get_positions(),
+            len(target),
+            target.get_atomic_numbers(),
+            target.get_positions(),
+            kmax,
+        )
+        new_pos = (target.get_positions() @ r.T) + t
+        target.set_positions(new_pos[p])
+        target.set_atomic_numbers(target.get_atomic_numbers()[p])
+    else:
+        minimize_rotation_and_translation(ref, target)
+    return target
+
+
+def align_path(frames, mode: AlignMode, use_ira=False, kmax=1.8):
+    """Applies the selected alignment strategy to the image sequence."""
+    if mode == AlignMode.NONE or len(frames) < 2:
         return frames
 
     ref = frames[0]
-    ref_pos = ref.get_positions()
-    ref_types = ref.get_atomic_numbers()
-    ref_natoms = len(ref)
 
-    aligned_frames = [ref.copy()]
+    if mode == AlignMode.ALL:
+        logging.info("Aligning [bold]all[/bold] images to reactant.")
+        return [ref.copy()] + [
+            perform_alignment(ref, f.copy(), use_ira, kmax) for f in frames[1:]
+        ]
 
-    # Initialize IRA if requested
-    ira_instance = None
-    if use_ira and ira_mod:
-        ira_instance = ira_mod.IRA()
-        logging.info(
-            "Using [bold magenta]IRA[/bold magenta] for optimal alignment and reordering."
-        )
+    if mode == AlignMode.ENDPOINTS:
+        logging.info("Aligning [bold]endpoints[/bold] (reactant and product) only.")
+        # Only the product (last frame) undergoes alignment relative to the reactant
+        aligned_product = perform_alignment(ref, frames[-1].copy(), use_ira, kmax)
+        # Intermediate frames remain unchanged in this specific mode logic,
+        # Usually, endpoint alignment implies ensuring the BCs match.
+        new_frames = [f.copy() for f in frames]
+        new_frames[-1] = aligned_product
+        return new_frames
 
-    for i in range(1, len(frames)):
-        current = frames[i].copy()
-
-        if ira_instance:
-            # Perform IRA match
-            r, t, p, _ = ira_instance.match(
-                ref_natoms,
-                ref_types,
-                ref_pos,
-                len(current),
-                current.get_atomic_numbers(),
-                current.get_positions(),
-                kmax,
-            )
-            # Apply rotation, translation, and permutation
[…]
