### 2025-12-23 dfb05ce
chore(pltneb): pass ira_kmax through
