### 2025-12-24 f881b5c
feat(consplit): fallback to ase minimize rotations
without permutations
