+++ rgpycrumbs/eon/con_splitter.py
@@ -57,23 +57,44 @@ class AlignMode(Enum):
 
 
 def perform_alignment(ref, target, use_ira=False, kmax=1.8):
-    """Performs the actual rotation/translation match between two structures."""
+    """
+    Performs the rotation/translation match between structures.
+
+    Attempts IRA-based reordering if requested. If the IRA library remains
+    unavailable or the matching process fails, the function falls back to
+    standard ASE rotation and translation minimization.
+    """
+
+    aligned_successfully = False
+
     if use_ira and ira_mod:
-        ira_instance = ira_mod.IRA()
-        r, t, p, _ = ira_instance.match(
-            len(ref),
-            ref.get_atomic_numbers(),
-            ref.get_positions(),
-            len(target),
-            target.get_atomic_numbers(),
-            target.get_positions(),
-            kmax,
-        )
-        new_pos = (target.get_positions() @ r.T) + t
-        target.set_positions(new_pos[p])
-        target.set_atomic_numbers(target.get_atomic_numbers()[p])
-    else:
+        try:
+            ira_instance = ira_mod.IRA()
+            r, t, p, _ = ira_instance.match(
+                len(ref),
+                ref.get_atomic_numbers(),
+                ref.get_positions(),
+                len(target),
+                target.get_atomic_numbers(),
+                target.get_positions(),
+                kmax,
+            )
+            # Apply transformation
+            new_pos = (target.get_positions() @ r.T) + t
+            target.set_positions(new_pos[p])
+            target.set_atomic_numbers(target.get_atomic_numbers()[p])
+            aligned_successfully = True
+            logging.info("IRA alignment and reordering complete.")
+        except Exception as e:
+            logging.warning(
+                f"IRA matching failed: {e}. Falling back to ASE Procrustes alignment."
+            )
+
+    if not aligned_successfully:
+        # Standard Procrustes alignment (no reordering)
         minimize_rotation_and_translation(ref, target)
+        logging.info("ASE rotation/translation minimization complete.")
+
     return target
 
 
