### 2025-12-24 86ee9a1
feat(pltneb): fallback to ase minimize rotations
