### 2025-12-31 c6b1b1c
maint(geom): refactor alignment out
