### 2025-12-31 e229bd5
chore(align): rework for clarity
