### 2025-12-31 e4dbcdf
chore(geom): even more classes for readability
