### 2025-12-31 67fb137
tst(align): add baseline tests
