### 2025-12-31 faea1b8
tst(align): add check for IRA behavior
