+++ tests/conftest.py
@@ -7,6 +7,7 @@ ENVIRONMENT_REQUIREMENTS = {
     "fragments": ["ase", "tblite"],
     "ptm": ["ase", "ovito"],
     "eon": ["ase", "eon"],
+    "align": ["ase"],
 }
 
 
