### 2025-12-31 9eac85b
chore(prj): setup markers more coherently
