### 2025-12-31 be03f15
chore(pixi): update environments
