### 2025-12-31 474a793
maint(consplit): Use structures and add docs
