+++ rgpycrumbs/eon/con_splitter.py
@@ -53,11 +53,18 @@ class AlignMode(Enum):
     """Defines structural alignment strategies."""
 
     NONE = "none"
-    ALL = "all"  # Align every frame to the reactant
-    ENDPOINTS = "endpoints"  # Align only reactant and product to each other
+    ALL = "all"
+    ENDPOINTS = "endpoints"
 
 
-def align_path(frames, mode: AlignMode, use_ira=False, kmax=1.8):
+class SplitMode(Enum):
+    """Defines trajectory validation strictness."""
+
+    NEB = "neb"  # Strict: must be a clean multiple of images_per_path
+    FLEX = "flex"  # Flexible: allows partial paths or simple slicing
+
+
+def align_path(frames, mode: AlignMode, iraconf: IRAConfig):
     """Applies the selected alignment strategy to the image sequence."""
     if mode == AlignMode.NONE or len(frames) < 2:
         return frames
@@ -65,18 +72,15 @@ def align_path(frames, mode: AlignMode, use_ira=False, kmax=1.8):
     ref = frames[0]
 
     if mode == AlignMode.ALL:
-        logging.info("Aligning [bold]all[/bold] images to reactant.")
+        logging.info("Aligning [bold]all[/bold] images to reactant reference.")
         return [ref.copy()] + [
             align_structure_robust(ref, f.copy(), IRAConfig(use_ira, kmax)).atoms
             for f in frames[1:]
         ]
 
     if mode == AlignMode.ENDPOINTS:
-        logging.info("Aligning [bold]endpoints[/bold] (reactant and product) only.")
-        # Only the product (last frame) undergoes alignment relative to the reactant
-        aligned_product = align_structure_robust(
-            ref, frames[-1].copy(), IRAConfig(use_ira, kmax)
-        ).atoms
+        logging.info("Aligning [bold]endpoints[/bold] (product to reactant) only.")
+        aligned_product = align_structure_robust(ref, frames[-1].copy(), iraconf).atoms
         # Intermediate frames remain unchanged in this specific mode logic,
         # Usually, endpoint alignment implies ensuring the BCs match.
         new_frames = [f.copy() for f in frames]
@@ -92,28 +96,34 @@ def align_path(frames, mode: AlignMode, use_ira=False, kmax=1.8):
     type=click.Path(exists=True, dir_okay=False, readable=True, path_type=Path),
 )
 @click.option(
+    "--mode",
+    type=click.Choice([m.value for m in SplitMode]),
+    default=SplitMode.FLEX.value,
+    help="Validation mode: 'neb' (strict multiples) or 'normal' (flexible).",
+)
+@click.option(
     "--output-dir",
     type=click.Path(file_okay=False, writable=True, path_type=Path),
     default=None,
-    help="Directory to save the output files.",
+    help="Directory to save output files. Defaults to the input filename stem.",
 )
 @click.option(
     "--images-per-path",
     type=int,
     required=True,
-    help="Number of images in a single NEB path.",
+    help="Number of images in a single NEB path (e.g., 7). [REQUIRED]",
 )
 @click.option(
     "--path-index",
     type=int,
     default=-1,
     show_default=True,
-    help="Index of the NEB path to extract. Use -1 for the last path.",
+    help="Index of the NEB path to extract (0-based). Use -1 for the last path.",
 )
 @click.option(
     "--center/--no-center",
     default=False,
-    help="Center the atoms in each frame around the origin.",
+    help="Center the atomic coordinates around the origin.",
 )
 @click.option(
     "--box-diagonal",
@@ -121,13 +131,13 @@ def align_path(frames, mode: AlignMode, use_ira=False, kmax=1.8):
     type=(float, float, float),
     default=(25.0, 25.0, 25.0),
     show_default=True,
-    help="Override the unit cell dimensions during processing.",
[…]
