### 2025-12-31 48da651
feat(pltneb): overlay multiple images
