### 2025-12-31 89ba8eb
feat(pltneb): take labels for additional files
