### 2025-12-31 76b3630
feat(pltneb): add saddle point estimate
