### 2025-12-31 ef77a8c
chore(pltneb): better defaults
