### 2025-12-31 af82882
chore(pltneb): take an empty string to default
