### 2025-12-31 83fb4b9
chore(pltneb): add a legend helper
