+++ pixi.toml
@@ -62,6 +62,7 @@ tblite-python = ">=0.5.0,<0.6"
 rgpycrumbs = { path = ".", editable=true }
 pytest = "*"
 click = ">=8.3.1, <9"
+numpy = ">=2.3.5, <3"
 
 [feature.ptm.dependencies]
 ovito = "*"
