### 2025-12-31 afb6d19
chore(pltneb): use better default labels
