### 2025-12-31 a5de49f
feat(pltneb): auto-determine smoothing for 2D
