### 2025-12-31 33f5a94
feat(pltneb): auto rbf smoother via global median
