+++ docs/newsfragments/+67aeedcf.eon.added.md
@@ -1,0 +1,1 @@
+feat(pltneb): Automatically determine the smoothing factor from the global median RMSD step distance
