### 2025-12-31 b1cb459
feat(pltneb): flip saddles and additional points
