### 2025-12-31 9f45e61
chore(pltneb): rework to squares
