### 2026-01-03 9fc7b93
chore(pltneb): pass ira_kmax consistently
