### 2026-01-07 0879ca2
bug(ira): rework to order things correctly
