+++ tests/test_align.py
@@ -46,12 +46,14 @@ def permuted_water(water_molecule):
     indices = [0, 2, 1]
     permuted = Atoms(
         symbols=[mobile.get_chemical_symbols()[i] for i in indices],
-        positions=mobile.get_positions()[indices],
+        positions=mobile.positions[indices],
     )
     return permuted
 
 
 class TestStructuralAlignment:
+    """Focuses on the physical consistency of atomic properties during permutation."""
+
     def test_identity_alignment(self, water_molecule):
         """Checks if alignment of a structure with itself returns correct status."""
         ref = water_molecule.copy()
@@ -61,7 +63,7 @@ class TestStructuralAlignment:
         result = align_structure_robust(ref, mobile, config)
 
         assert result.method == AlignmentMethod.ASE_PROCRUSTES
-        assert np.allclose(mobile.get_positions(), ref.get_positions())
+        assert np.allclose(mobile.positions, ref.positions)
 
     def test_ase_fallback_rotation(self, water_molecule, rotated_water):
         """Verifies that ASE Procrustes handles rotation when IRA is disabled."""
@@ -70,9 +72,7 @@ class TestStructuralAlignment:
         result = align_structure_robust(water_molecule, rotated_water, config)
 
         assert result.method == AlignmentMethod.ASE_PROCRUSTES
-        assert np.allclose(
-            rotated_water.get_positions(), water_molecule.get_positions(), atol=1e-5
-        )
+        assert np.allclose(rotated_water.positions, water_molecule.positions, atol=1e-5)
 
     @requires_ira
     def test_ira_permutation_success(self, water_molecule, permuted_water):
@@ -84,9 +84,7 @@ class TestStructuralAlignment:
         )
         assert result.method == AlignmentMethod.IRA_PERMUTATION
         # The permuted water should now match the reference positions AND atomic order
-        assert np.allclose(
-            permuted_water.get_positions(), water_molecule.get_positions()
-        )
+        assert np.allclose(permuted_water.positions, water_molecule.positions)
         assert list(permuted_water.get_chemical_symbols()) == list(
             water_molecule.get_chemical_symbols()
         )
@@ -104,9 +102,7 @@ class TestStructuralAlignment:
         result = align_structure_robust(water_molecule, rotated_water, config)
 
         assert result.method == AlignmentMethod.ASE_PROCRUSTES
-        assert np.allclose(
-            rotated_water.get_positions(), water_molecule.get_positions(), atol=1e-5
-        )
+        assert np.allclose(rotated_water.positions, water_molecule.positions, atol=1e-5)
 
     @requires_ira
     def test_ase_fails_on_permutation_but_ira_succeeds(self, water_molecule):
@@ -114,7 +110,7 @@ class TestStructuralAlignment:
         indices = [0, 2, 1]
         permuted_water = Atoms(
             symbols=[water_molecule.get_chemical_symbols()[i] for i in indices],
-            positions=water_molecule.get_positions()[indices],
+            positions=water_molecule.positions[indices],
         )
 
         # Break the C2v symmetry by slightly nudging one hydrogen atom.
@@ -148,3 +144,94 @@ class TestStructuralAlignment:
         # IRA should still produce a lower error because it correctly identifies
         # which hydrogen corresponds to the reference positions.
         assert ira_dist < ase_dist
+
+    @requires_ira
+    def test_mass_and_identity_sync(self, water_molecule):
+        """
+        Ensures that masses and identities follow the permutation.
+        If we align a deuterated water where the D and H are swapped,
+        the resulting structure must have the D at the correct reference index.
+        """
[…]
