### 2026-01-07 927ea95
chore(pixi): add to the dev env
