### 2026-01-17 76a89ed
feat(pltneb): use a strip below the main plot
