+++ rgpycrumbs/eon/plt_neb.py
@@ -29,6 +29,7 @@ https://realpython.com/python-script-structure/
 #   "matplotlib",
 #   "numpy",
 #   "scipy",
+#   "adjustText",
 #   "cmcrameri",
 #   "rich",
 #   "ase",
@@ -57,6 +58,7 @@ from ase.io import read as ase_read
 from ase.io import write as ase_write
 from matplotlib.collections import LineCollection
 from matplotlib.colors import LinearSegmentedColormap
+from matplotlib.gridspec import GridSpec
 from matplotlib.offsetbox import AnnotationBbox, OffsetImage
 from matplotlib.patches import ArrowStyle
 from rich.logging import RichHandler
@@ -68,6 +70,7 @@ from scipy.interpolate import (
     splrep,
 )
 from scipy.signal import savgol_filter
+from adjustText import adjust_text
 
 from rgpycrumbs._aux import _import_from_parent_env
 from rgpycrumbs.geom.api.alignment import IRAConfig, align_structure_robust
@@ -175,7 +178,6 @@ RUHI_THEME = PlotTheme(
     highlight_color="black",
 )
 
-# Updated THEMES dictionary
 THEMES = {
     "cmc.batlow": BATLOW_THEME,
     "ruhi": RUHI_THEME,
@@ -408,24 +410,24 @@ def _load_or_compute_data(
     :return: The requested data.
     :rtype: pl.DataFrame
     """
-    # 1. Attempt to load from cache
+    # Attempt to load from cache
     if cache_file and cache_file.exists() and not force_recompute:
         log.info(
             f"Loading cached {context_name} data from [green]{cache_file}[/green]..."
         )
         try:
             df = pl.read_parquet(cache_file)
-            validation_check(df)  # Will raise ValueError if invalid
+            validation_check(df)
             log.info(f"Loaded {df.height} rows from cache.")
             return df
         except Exception as e:
             log.warning(f"Cache load failed or invalid: {e}. Recomputing...")
 
-    # 2. Compute if cache failed, didn't exist, or recompute requested
+    # Compute if cache failed, didn't exist, or recompute requested
     log.info(f"Computing {context_name} data...")
     df = computation_callback()
 
-    # 3. Save to cache
+    # Save to cache
     if cache_file:
         log.info(f"Saving {context_name} cache to [green]{cache_file}[/green]...")
         try:
@@ -434,6 +436,66 @@ def _load_or_compute_data(
             log.error(f"Failed to write cache file: {e}")
 
     return df
+
+
+def plot_structure_strip(
+    ax,
+    atoms_list,
+    labels=None,
+    zoom=0.3,
+    ase_rotation="0x, 90y, 0z",
+    theme_color="black",
+):
+    """
+    Renders a horizontal gallery of atomic structures (equidistant).
+    """
+    # completely turn off axis lines/ticks
+    ax.axis("off")
+
+    n_plot = len(atoms_list)
+    # Set limits: 0 to N-1 with padding
+    ax.set_xlim(-0.5, n_plot - 0.5)
+    ax.set_ylim(-0.5, 0.5)
+
+    for i in range(n_plot):
+        atoms = atoms_list[i]
+        # Force integer positioning (0, 1, 2...)
+        x_pos = i
+
+        # 1. High-Res Rendering
+        buf = io.BytesIO()
+        ase_write(
+            buf, atoms, format="png", rotation=ase_rotation, show_unit_cell=0, scale=100
+        )
+        buf.seek(0)
+        img_data = plt.imread(buf)
+        buf.close()
+
+        # 2. Adjust Zoom
+        effective_zoom = zoom * 0.35
+        imagebox = OffsetImage(img_data, zoom=effective_zoom)
+
+        ab = AnnotationBbox(
[…]
