### 2026-01-17 c50a8f3
feat(pltneb): adjust number of rows dynamically
