### 2026-01-17 53cb677
chore(pltneb): use the older strategy for 1D
