### 2026-01-28 28ac681
feat(eon): add dictionary support
for generating configurations
