### 2026-01-28 f791c05
doc(rel): add changelog entries
