+++ docs/newsfragments/28ac681.eon.added.md
@@ -1,0 +1,1 @@
+feat(eon): add dictionary support
