+++ docs/newsfragments/a60da70.jup.added.md
@@ -1,0 +1,1 @@
+feat(jup): add subprocess run helpers for atomistic-cookbook
