### 2026-01-28 bb24e4d
rel(0.1.0): add changelog
