+++ docs/newsfragments/+6e570f67.geom.added.md
@@ -1,1 +1,0 @@
-Added robust alignment API using Isomorphic Robust Alignment (IRA) with ASE fallback.
