+++ docs/newsfragments/28ac681.eon.added.md
@@ -1,1 +1,0 @@
-feat(eon): add dictionary support
