+++ docs/newsfragments/+f1211da2.consplit.added.md
@@ -1,1 +1,0 @@
-Added alignment using IRA for splitting con files.
