+++ towncrier.toml
@@ -3,8 +3,8 @@ directory = "docs/newsfragments"
 filename = "CHANGELOG.md"
 start_string = "<!-- towncrier release notes start -->\n"
 underlines = ["", "", ""]
-title_format = "## [{version}](https://github.com/theochemui/eongit/tree/{version}) - {project_date}"
-issue_format = "[#{issue}](https://github.com/theochemui/eongit/issues/{issue})"
+title_format = "## [{version}](https://github.com/HaoZeke/rgpycrumbs/tree/{version}) - {project_date}"
+issue_format = "[#{issue}](https://github.com/HaoZeke/rgpycrumbs/issues/{issue})"
 
 [[tool.towncrier.type]]
 directory = "security"
