### 2026-01-28 5596508
chore(jupyter): cleanup slightly overactive cmd
