+++ rgpycrumbs/run/jupyter.py
@@ -90,64 +90,3 @@ def run_command_or_exit(
     except subprocess.TimeoutExpired:
         print("Command timed out", file=sys.stderr)
         sys.exit(124)
-
-
-def get_neb_plot_cmd(
-    mode: str,
-    con_file: str = "neb.con",
-    output_file: str = "plot.png",
-    title: str = "",
-    theme: str = "ruhi",
-    rotation: str = "90x,0y,0z",
-) -> list[str]:
-    """
-    Constructs the CLI command for rgpycrumbs plotting to avoid clutter in notebooks.
-    mode: 'profile' (1D) or 'landscape' (2D)
-    """
-    base_cmd = [
-        sys.executable,
-        "-m",
-        "rgpycrumbs.cli",
-        "eon",
-        "plt_neb",
-        "--theme",
-        theme,
-        "--con-file",
-        con_file,
-        "--output-file",
-        output_file,
-        "--ase-rotation",
-        rotation,
-        "--facecolor",
-        "white",
-        "--plot-structures",
-        "crit_points",
-    ]
-
-    if title:
-        base_cmd.extend(["--title", title])
-
-    if mode == "profile":
-        base_cmd.extend(["--plot-type", "profile"])
-    elif mode == "landscape":
-        base_cmd.extend(
-            [
-                "--plot-type",
-                "landscape",
-                "--rc-mode",
-                "path",
-                "--fontsize-base",
-                "16",
-                "--landscape-mode",
-                "surface",
-                "--landscape-path",
-                "all",
-                "--show-pts",
-                "--surface-type",
-                "rbf",
-            ]
-        )
-    else:
-        raise ValueError(f"Unknown plot mode: {mode}")
-
-    return base_cmd
