+++ pixi.toml
@@ -17,6 +17,10 @@ pytest = ">=8.4.2,<9"
 [feature.docs.dependencies]
 emacs = ">=30.2,<31"
 nodejs = ">=25.2.1,<25.3"
+sphinx-favicon = ">=1.0.1,<2"
+ase = ">=3.27.0,<4"
+numpy = ">=2.4.1,<3"
+mlflow = ">=3.8.1,<4"
 
 [feature.docs.pypi-dependencies]
 shibuya = ">=2025.12.13, <2026"
