+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Changelog
+
 ## [v0.1.0](https://github.com/HaoZeke/rgpycrumbs/tree/v0.1.0) - 2026-01-28
 
 ### Added
