### 2026-01-28 45167c7
chore(pixi): add more deps
