+++ pixi.toml
@@ -23,6 +23,7 @@ numpy = ">=2.4.1,<3"
 mlflow = ">=3.8.1,<4"
 
 [feature.docs.pypi-dependencies]
+rgpycrumbs = { path = ".", editable = true }
 shibuya = ">=2025.12.13, <2026"
 sphinx-click = ">=6.2.0, <7"
 sphinx = ">=9.0.4, <10"
