### 2026-01-28 6d589b2
chore(crier): switch changelog format
