+++ pixi.toml
@@ -21,6 +21,7 @@ sphinx-favicon = ">=1.0.1,<2"
 ase = ">=3.27.0,<4"
 numpy = ">=2.4.1,<3"
 mlflow = ">=3.8.1,<4"
+adjusttext = ">=1.3.0,<2"
 
 [feature.docs.pypi-dependencies]
 rgpycrumbs = { path = ".", editable = true }
@@ -32,7 +33,6 @@ sphinx-autoapi = ">=3.6.0, <4"
 sphobjinv = ">=2.3.1.3, <3"
 autodoc-pydantic = ">=2.2.0, <3"
 sphinx-sitemap = ">=2.9.0, <3"
-pycmark = ">=0.9.6, <0.10"
 
 [environments]
 docs = {features = ["docs"]}
