+++ _sources/tools/eon/index.rst.txt
@@ -11,6 +11,7 @@ These tools work with `EON output files <https://eondocs.org>`_.
 
    con_splitter
    generate_nwchem_input
+   mlflow
    plt_neb
    ptmdisp
 
