+++ _sources/autoapi/rgpycrumbs/eon/index.rst.txt
@@ -13,6 +13,7 @@ Submodules
    /autoapi/rgpycrumbs/eon/_mlflow/index
    /autoapi/rgpycrumbs/eon/con_splitter/index
    /autoapi/rgpycrumbs/eon/generate_nwchem_input/index
+   /autoapi/rgpycrumbs/eon/helpers/index
    /autoapi/rgpycrumbs/eon/plt_neb/index
    /autoapi/rgpycrumbs/eon/ptmdisp/index
 
