+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+rgpycrumbs v0.1.2 (2026-01-28)
+==============================
+
+Fixed
+-----
+
+- Additional imports
+
+
 rgpycrumbs v0.1.1 (2026-01-28)
 ==============================
 
