### 2026-02-05 #275 @kysolvik
Hi @HaoZeke! 👋

I'm a member of our EiC team. Sorry for the delay in response, we had turnover at the new year and are just catching up now.

Thanks for your presubmission inquiry regarding rgpycrumbs! It's a neat package that seems to contain a lot of potentially useful tools, and the isolated run-time environments using PEP 723 metadata standards are elegant. Unfortunately, as it's structured now we think that the functionality is too loosely connected for pyOpenSci since it contains various tools that were used on a per-paper basis that might not be generally applicable for other users (please correct me if I'm wrong!). Would you consider splitting off a portion of the package, say the eOn helpers, into a smaller package with more general utility?

Happy to chat more and answer any questions you have!
