### 2026-02-05 #275 @kysolvik
Note that I closed the issue for now just for book-keeping but happy to keep chatting!
