### 2026-02-06 #275 @HaoZeke
Hi @kysolvik, I did think about spinning them out into separate packages, however, I personally feel like that'd be a waste. The concept here (and what I understood to be part of the goals of PyOpenSci) is to have tools which are "correct" in the sense of good design / programming and do the "right thing" scientifically.

One of the biggest issues with the landscape today is the "poisoning" of dependencies to use existing tools. For instance, take any of the examples, say plotting the NEB:

This currently involves the PEP 723 launcher, so any user can incorporate it into their workflow, without having to worry about, say, which ASE version was used to load the files, and if said version conflicts with their own installed packages. In particular, this means even 10 years down the line someone can use the package (which itself is of course versioned) reproducibly, without trying to re-haul dependencies just to be able to use the tools within the `rgpycrumbs` package.

To me, it is this design makes sense as a reviewable artifact for the PyOpenSci ecosystem, does that make any sense?

My concern is that maintaining packages involves a lot of overhead (documentation, examples, and dependencies) and while it is possible to have several (I certainly have a lot), a more principled way would be to enable a "loosely-coupled" but unified set of tools, like the package submitted (or SciPy for that matter, which also is similar in that the explicit dependencies are minimal, but several modules are basically independent and maintained by separate groups).
