### 2026-02-07 559a455
feat(pltneb): refactor to use chemparseplot
