+++ rgpycrumbs/eon/plt_neb.py
@@ -34,50 +34,46 @@ https://realpython.com/python-script-structure/
 #   "rich",
 #   "ase",
 #   "polars",
+#   "chemparseplot @ file:///home/rgoswami/Git/Github/Python/chemparseplot/",
+#   "rgpycrumbs",
 # ]
 # ///
 
-import glob
-import io
 import logging
-import re
-import sys
-from collections import namedtuple
-from collections.abc import Callable
-from dataclasses import dataclass, replace
-from enum import Enum
 from pathlib import Path
 
 import click
-import cmcrameri.cm  # noqa: F401
-import matplotlib as mpl
-import matplotlib.patheffects
 import matplotlib.pyplot as plt
+import matplotlib.patheffects as path_effects
 import numpy as np
 import polars as pl
-from adjustText import adjust_text
-from ase.io import read as ase_read
-from ase.io import write as ase_write
-from matplotlib.collections import LineCollection
-from matplotlib.colors import LinearSegmentedColormap
 from matplotlib.gridspec import GridSpec
-from matplotlib.offsetbox import AnnotationBbox, OffsetImage
 from matplotlib.patches import ArrowStyle
 from rich.logging import RichHandler
-from scipy.interpolate import (
-    CubicHermiteSpline,
-    RBFInterpolator,
-    griddata,
-    splev,
-    splrep,
+
+from adjustText import adjust_text
+
+# --- Library Imports ---
+from chemparseplot.parse.file_ import find_file_paths
+from chemparseplot.parse.eon.neb import (
+    aggregate_neb_landscape_data,
+    compute_profile_rmsd,
+    load_structures_and_calculate_additional_rmsd,
+    estimate_rbf_smoothing,
 )
-from scipy.signal import savgol_filter
-
-from rgpycrumbs._aux import _import_from_parent_env
-from rgpycrumbs.geom.api.alignment import IRAConfig, align_structure_robust
-
-# IRA is optional, use None if not present
-ira_mod = _import_from_parent_env("ira_mod")
+from chemparseplot.plot.neb import (
+    InsetImagePos,
+    plot_energy_path,
+    plot_landscape_path_overlay,
+    plot_landscape_surface,
+    plot_structure_inset,
+    plot_structure_strip,
+)
+from chemparseplot.plot.theme import (
+    apply_axis_theme,
+    get_theme,
+    setup_global_theme,
+)
 
 # --- Logging Setup ---
 logging.basicConfig(
@@ -88,928 +84,11 @@ logging.basicConfig(
 log = logging.getLogger("rich")
 
 
-# --- Enumerations, Constants, and Themes ---
-
-# --- Color Definitions ---
-RUHI_COLORS = {
-    "coral": "#FF655D",
-    "sunshine": "#F1DB4B",
-    "teal": "#004D40",
-    "sky": "#1E88E5",
-    "magenta": "#D81B60",
-}
-
-MIN_PATH_LENGTH = 1e-6
-
-
-# --- Custom Colormap Generation ---
-def build_cmap(hex_list, name):
-    """Build and register a LinearSegmentedColormap from a list of hex colors."""
-    cols = [c.strip() for c in hex_list]
-    cmap = LinearSegmentedColormap.from_list(name, cols, N=256)
-    mpl.colormaps.register(cmap)
-    return cmap
-
-
-# Build and register Ruhi colormaps
-build_cmap(
-    [
-        RUHI_COLORS["coral"],
-        RUHI_COLORS["sunshine"],
-        RUHI_COLORS["teal"],
-        RUHI_COLORS["sky"],
-        RUHI_COLORS["magenta"],
-    ],
-    name="ruhi_full",
-)
[…]
