### 2026-02-07 7ed7573
feat(pltneb): bring back additional-con settings
