+++ rgpycrumbs/eon/plt_neb.py
@@ -43,6 +43,7 @@ import logging
 from pathlib import Path
 
 import click
+import matplotlib as mpl
 import matplotlib.pyplot as plt
 import matplotlib.patheffects as path_effects
 import numpy as np
@@ -412,7 +413,6 @@ def main(
     )
     setup_global_theme(active_theme)
 
-    # 2. Setup Figure
     if fig_height and aspect_ratio:
         figsize = (fig_height * aspect_ratio, fig_height)
     elif fig_height or aspect_ratio:
@@ -444,7 +444,6 @@ def main(
 
     apply_axis_theme(ax, active_theme)
 
-    # 3. Data Loading (Delegated to Library)
     atoms_list = None
     additional_atoms_data = []
 
@@ -463,7 +462,6 @@ def main(
                 log.critical("Cannot proceed without structures. Exiting.")
                 exit(1)
 
-    # 4. Plot Execution
     if plot_type == "landscape":
         # --- Landscape Plot ---
         dat_paths = find_file_paths(input_dat_pattern)
@@ -500,6 +498,8 @@ def main(
             r_all = df_surface["r"].to_numpy()
             p_all = df_surface["p"].to_numpy()
             z_all = df_surface["z"].to_numpy()
+            gr_all = df_surface["grad_r"].to_numpy()
+            gp_all = df_surface["grad_p"].to_numpy()
 
             # Heuristic for RBF smoothing if missing
             if surface_type == "rbf" and rbf_smoothing is None:
@@ -510,6 +510,8 @@ def main(
                 ax,
                 r_all,
                 p_all,
+                gr_all,
+                gp_all,
                 z_all,
                 method=surface_type,
                 rbf_smooth=rbf_smoothing,
@@ -546,24 +548,40 @@ def main(
             label="SP",
         )
 
-        # Structure Strip / Insets
-        if has_strip and atoms_list:
-            # Determine indices
-            if plot_structures == "all":
-                indices = list(range(len(atoms_list)))
-            else:
-                indices = sorted(list({0, saddle_idx, len(atoms_list) - 1}))
+        if additional_atoms_data:
+            marker_cmap = mpl.colormaps.get_cmap("tab10")
+            for i, (_, add_r, add_p, add_label) in enumerate(additional_atoms_data):
+                color = marker_cmap(i % 10)
+                ax.plot(
+                    add_r,
+                    add_p,
+                    marker="*",
+                    markersize=int(active_theme.font_size * 1.1),
+                    color=color,
+                    markeredgecolor="white",
+                    markeredgewidth=1.0,
+                    linestyle="None",
+                    zorder=102,
+                    label=add_label,
+                )
 
-            # Build payload
+        if has_strip and atoms_list:
+            indices = (
+                list(range(len(atoms_list)))
+                if plot_structures == "all"
+                else sorted(list({0, saddle_idx, len(atoms_list) - 1}))
+            )
             strip_payload = []
             for i in indices:
-                lbl = str(i)
-                if i == 0:
-                    lbl = "R"
-                elif i == saddle_idx:
-                    lbl = "SP"
-                elif i == len(atoms_list) - 1:
-                    lbl = "P"
+                lbl = (
+                    "R"
+                    if i == 0
+                    else (
+                        "SP"
+                        if i == saddle_idx
+                        else ("P" if i == len(atoms_list) - 1 else str(i))
+                    )
+                )
                 strip_payload.append(
                     {
                         "atoms": atoms_list[i],
@@ -692,45 +710,68 @@ def main(
             )
 
             if highlight_last and is_last and atoms_list and plot_structures != "none":
-                indices = []
-                if plot_structures == "all":
-                    indices = list(range(len(atoms_list)))
-                elif plot_structures == "crit_points":
-                    y_vals = data[y_col]
[…]
