### 2026-02-07 4ae12de
feat(pltneb): take in eOn's new sp.con
