### 2026-02-07 d8af5cd
feat(pltneb): enable augmenting data for neb 2D
