### 2026-02-07 d1d2a84
feat(pltneb): cache augmented data
