### 2026-02-09 9481702
chore(pltneb): handle more options
and pin chemparseplot
