+++ _sources/autoapi/rgpycrumbs/eon/plt_neb/index.rst.txt
@@ -33,31 +33,11 @@ Attributes
 
 .. autoapisummary::
 
-   rgpycrumbs.eon.plt_neb.ira_mod
    rgpycrumbs.eon.plt_neb.log
-   rgpycrumbs.eon.plt_neb.RUHI_COLORS
-   rgpycrumbs.eon.plt_neb.MIN_PATH_LENGTH
-   rgpycrumbs.eon.plt_neb.BATLOW_THEME
-   rgpycrumbs.eon.plt_neb.RUHI_THEME
-   rgpycrumbs.eon.plt_neb.THEMES
    rgpycrumbs.eon.plt_neb.DEFAULT_INPUT_PATTERN
    rgpycrumbs.eon.plt_neb.DEFAULT_PATH_PATTERN
    rgpycrumbs.eon.plt_neb.ROUNDING_DF
    rgpycrumbs.eon.plt_neb.IRA_KMAX_DEFAULT
-
-
-Classes
--------
-
-.. autoapisummary::
-
-   rgpycrumbs.eon.plt_neb.PlotTheme
-   rgpycrumbs.eon.plt_neb.PlotType
-   rgpycrumbs.eon.plt_neb.RCMode
-   rgpycrumbs.eon.plt_neb.PlotMode
-   rgpycrumbs.eon.plt_neb.SplineMethod
-   rgpycrumbs.eon.plt_neb.SmoothingParams
-   rgpycrumbs.eon.plt_neb.InsetImagePos
 
 
 Functions
@@ -65,179 +45,13 @@ Functions
 
 .. autoapisummary::
 
-   rgpycrumbs.eon.plt_neb.build_cmap
-   rgpycrumbs.eon.plt_neb.setup_global_theme
-   rgpycrumbs.eon.plt_neb.apply_plot_theme
-   rgpycrumbs.eon.plt_neb.load_paths
-   rgpycrumbs.eon.plt_neb.calculate_rmsd_from_ref
-   rgpycrumbs.eon.plt_neb.calculate_landscape_coords
-   rgpycrumbs.eon.plt_neb._load_or_compute_data
-   rgpycrumbs.eon.plt_neb.plot_structure_strip
-   rgpycrumbs.eon.plt_neb.plot_single_inset
-   rgpycrumbs.eon.plt_neb.plot_structure_insets
-   rgpycrumbs.eon.plt_neb.plot_energy_path
-   rgpycrumbs.eon.plt_neb.plot_eigenvalue_path
-   rgpycrumbs.eon.plt_neb.plot_landscape_path
-   rgpycrumbs.eon.plt_neb.plot_interpolated_grid
-   rgpycrumbs.eon.plt_neb.plot_interpolated_rbf
-   rgpycrumbs.eon.plt_neb.setup_plot_aesthetics
    rgpycrumbs.eon.plt_neb.main
-   rgpycrumbs.eon.plt_neb._setup_theme
-   rgpycrumbs.eon.plt_neb._run_dependency_checks
-   rgpycrumbs.eon.plt_neb._determine_figsize
-   rgpycrumbs.eon.plt_neb._load_structures
-   rgpycrumbs.eon.plt_neb._validate_data_atoms_match
-   rgpycrumbs.eon.plt_neb._get_profile_labels
-   rgpycrumbs.eon.plt_neb._aggregate_all_paths
-   rgpycrumbs.eon.plt_neb._plot_landscape
-   rgpycrumbs.eon.plt_neb._plot_profile
 
 
 Module Contents
 ---------------
 
-.. py:data:: ira_mod
-
 .. py:data:: log
-
-.. py:data:: RUHI_COLORS
-
-.. py:data:: MIN_PATH_LENGTH
-   :value: 1e-06
-
-
[…]
