### 2026-02-09 #275 @kysolvik
Hi @HaoZeke! Thanks for taking the time to write back! I'll bring this back to the rest of the editorial team to discuss more, since I do think this is a well-designed and well-written package. For now, I'd say the most important factor is how generally applicable the various scripts are to other scientists' workflows.

My initial impression (especially based on your pre-submission inquiry and the repo's name and README) was that the package contains scripts that were used for specific papers and your dissertation, but may not be as generally applicable to others beyond reproducing those specific workflows. That's why I asked about spinning off subsets into their own packages, since my thought was there might be a slice of the functionality that is more generally useful (e.g. the plotting helpers). That said, I very well could have misunderstood some of the functionality, this is outside of my field (although coincidentally my undergrad background is in computational chemistry).

Leveraging PEP 723 is a clever and clean way of combining scripts that need various, possibly conflicting dependencies, but by itself I don't think that makes it in-scope for us if the package doesn't generally support other scientists' workflows and make sense as a cohesive package.

Do you think all of these sub-packages would be used by other researchers? Do you think wrapping the various functionalities would help or hinder discoverability (e.g. would someone looking for one sub-package likely want to use another)? Do you have examples of other researchers using the package, even just within your research group?

None of this is a reflection of the quality of the package, we're just trying to determine if pyOpenSci is the best place for it.

Reopening the issue so others can find it. Thanks again for your thoughtful response!
