### 2026-02-16 8f70a77
feat: add library modules from chemparseplot
Move computational modules from chemparseplot into rgpycrumbs as
importable library modules: surfaces (JAX-based kernel fitting),
interpolation (spline utilities), basetypes (shared data structures),
geom/analysis (structure analysis), and geom/ira (IRA matching).
