+++ rgpycrumbs/surfaces.py
@@ -1,0 +1,896 @@
+import jax
+import jax.numpy as jnp
+import jax.scipy.optimize as jopt
+from jax import jit, vmap
+
+# Force float32 for speed/viz
+jax.config.update("jax_enable_x64", False)
+
+# ==============================================================================
+# HELPER: GENERIC LOSS FUNCTIONS
+# ==============================================================================
+
+
+def safe_cholesky_solve(K, y, noise_scalar, jitter_steps=3):
+    """Retries Cholesky with increasing jitter if it fails."""
+    N = K.shape[0]
+
+    # Try successively larger jitters: 1e-6, 1e-5, 1e-4
+    for i in range(jitter_steps):
+        jitter = (noise_scalar + 10 ** (-6 + i)) * jnp.eye(N)
+        try:
+            L = jnp.linalg.cholesky(K + jitter)
+            alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
+            log_det = 2.0 * jnp.sum(jnp.log(jnp.diag(L)))
+            return alpha, log_det
+        except:
+            continue
+
+    # Fallback for compilation safety (NaN propagation)
+    return jnp.zeros_like(y), jnp.nan
+
+
+def generic_negative_mll(K, y, noise_scalar):
+    alpha, log_det = safe_cholesky_solve(K, y, noise_scalar)
+
+    data_fit = 0.5 * jnp.dot(y.flatten(), alpha.flatten())
+    complexity = 0.5 * log_det
+
+    cost = data_fit + complexity
+    # heavy penalty if Cholesky failed (NaN)
+    return jnp.where(jnp.isnan(cost), 1e9, cost)
+
+
+# ==============================================================================
+# 1. TPS IMPLEMENTATION (Optimizable)
+# ==============================================================================
+
+
+@jit
+def _tps_kernel_matrix(x):
+    d2 = jnp.sum((x[:, None, :] - x[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K = r**2 * jnp.log(r)
+    return K
+
+
+def negative_mll_tps(log_params, x, y):
+    # TPS only really has a smoothing parameter to tune in this context
+    # (Length scale is inherent to the radial basis).
+    smoothing = jnp.exp(log_params[0])
+    K = _tps_kernel_matrix(x)
+    return generic_negative_mll(K, y, smoothing)
+
+
+@jit
+def _tps_solve(x, y, sm):
+    K = _tps_kernel_matrix(x)
+    K = K + jnp.eye(x.shape[0]) * sm
+
+    # Polynomial Matrix
+    N = x.shape[0]
+    P = jnp.concatenate([jnp.ones((N, 1), dtype=jnp.float32), x], axis=1)
+    M = P.shape[1]
+
+    # Solve System
+    zeros = jnp.zeros((M, M), dtype=jnp.float32)
+    top = jnp.concatenate([K, P], axis=1)
+    bot = jnp.concatenate([P.T, zeros], axis=1)
+    lhs = jnp.concatenate([top, bot], axis=0)
+    rhs = jnp.concatenate([y, jnp.zeros(M, dtype=jnp.float32)])
+
+    coeffs = jnp.linalg.solve(lhs, rhs)
+    return coeffs[:N], coeffs[N:]
+
+
+@jit
+def _tps_predict(x_query, x_obs, w, v):
+    d2 = jnp.sum((x_query[:, None, :] - x_obs[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K_q = r**2 * jnp.log(r)
+
+    P_q = jnp.concatenate(
[…]
