### 2026-02-16 1d65f73
feat: add public API, click CLI, and optional dep groups
- Add lazy imports via __getattr__ in __init__.py for surfaces,
  basetypes, and interpolation modules
- Convert CLI from argparse to click groups while keeping uv dispatch
- Add optional dependency groups: surfaces, analysis, interpolation, ira
- Add [project.scripts] entry point for rgpycrumbs CLI
- Update pixi.toml with surfaces environment and test deps
