+++ pyproject.toml
@@ -29,9 +29,9 @@ test        = "pytest --cov=rgpycrumbs tests"
 
 [project]
 name = "rgpycrumbs"
-description = "A dispatcher-based analytical and visualization suite for chemical physics"
+description = "A dispatcher-based analytical and computational suite for chemical physics"
 readme = "README.md"
-keywords = [ "plots", "parsers", "analysis" ]
+keywords = [ "compchem", "analysis", "surfaces", "interpolation" ]
 license = { text = "MIT" }
 authors = [ { name = "Rohit Goswami", email = "rgoswami@ieee.org" } ]
 requires-python = ">=3.10"
@@ -39,13 +39,26 @@ classifiers = [
   "Development Status :: 4 - Beta",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: Implementation :: CPython",
 ]
+dependencies = [
+  "click>=8.1",
+  "numpy>=1.24",
+]
 dynamic = [ "version" ]
+
+[project.optional-dependencies]
+surfaces = [ "jax>=0.4" ]
+analysis = [ "ase>=3.22", "scipy>=1.11" ]
+interpolation = [ "scipy>=1.11" ]
+ira = [ "ira_mod" ]
+
+[project.scripts]
+rgpycrumbs = "rgpycrumbs.cli:main"
+
 [project.urls]
 Documentation = "https://github.com/HaoZeke/rgpycrumbs#readme"
 Issues        = "https://github.com/HaoZeke/rgpycrumbs/issues"
@@ -145,5 +158,7 @@ markers = [
     "ptm: mark a test as needing only the ptm environment (run with '-m ptm').",
     "eon: mark a test as needing only the eon environment (run with '-m eon').",
     "align: mark a test as needing only the align environment (run with '-m align').",
+    "surfaces: mark a test as needing jax (run with '-m surfaces').",
+    "ira: mark a test as needing ira_mod (run with '-m ira').",
     "slow: mark tests as slow (deselect with '-m \"not slow\"').",
 ]
