+++ pixi.toml
@@ -39,6 +39,7 @@ docs = {features = ["docs"]}
 test = { features = ["test"] }
 fragments = { features = ["fragments", "test"] }
 ptm = { features = ["ptm", "test"] }
+surfaces = { features = ["surfaces", "test"] }
 eonmlflow = { features = ["eonmlflow", "test"] }
 
 [pypi-options]
@@ -69,6 +70,13 @@ rgpycrumbs = { path = ".", editable=true }
 pytest = "*"
 click = ">=8.3.1, <9"
 numpy = ">=2.3.5, <3"
+scipy = ">=1.11"
+
+[feature.test.dependencies]
+ase = ">=3.23,<4"
+
+[feature.surfaces.pypi-dependencies]
+jax = ">=0.4"
 
 [feature.ptm.dependencies]
 ovito = "*"
