### 2026-02-16 31d12ed
test: add tests for new library modules
- test_surfaces: 24 tests covering kernel matrices, model fit/predict,
  optimization, batched prediction, and factory function
- test_analysis: 14 tests for structure analysis including edge cases,
  matrix properties, fragment detection, and corrected distances
- test_ira: 9 tests for IRA dataclass, error handling, matching, and RMSD
- test_basetypes: 8 tests for all shared data structures
- test_interpolation: 2 tests for spline roundtrip
- Update conftest.py with surfaces and ira markers
