+++ tests/test_analysis.py
@@ -1,0 +1,201 @@
+import pytest
+
+from tests.conftest import skip_if_not_env
+
+skip_if_not_env("align")
+
+import numpy as np  # noqa: E402
+from ase.atoms import Atoms  # noqa: E402
+
+from rgpycrumbs.geom.analysis import analyze_structure  # noqa: E402
+
+pytestmark = pytest.mark.align
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def water_molecule():
+    """Creates a single water molecule."""
+    return Atoms("H2O", positions=[[0, 0, 0], [0, 0.76, 0.59], [0, -0.76, 0.59]])
+
+
+@pytest.fixture
+def water_dimer():
+    """Creates two water molecules separated by 5 Angstroms."""
+    h2o = Atoms("H2O", positions=[[0, 0, 0], [0, 0.76, 0.59], [0, -0.76, 0.59]])
+    dimer = h2o.copy()
+    h2o_2 = h2o.copy()
+    h2o_2.translate([5.0, 0, 0])
+    dimer.extend(h2o_2)
+    return dimer
+
+
+@pytest.fixture
+def single_atom():
+    """A single hydrogen atom."""
+    return Atoms("H", positions=[[0, 0, 0]])
+
+
+@pytest.fixture
+def h2_molecule():
+    """A simple H2 molecule with a typical bond length."""
+    return Atoms("H2", positions=[[0, 0, 0], [0, 0, 0.74]])
+
+
+@pytest.fixture
+def three_fragments():
+    """Three water molecules well-separated."""
+    h2o = Atoms("H2O", positions=[[0, 0, 0], [0, 0.76, 0.59], [0, -0.76, 0.59]])
+    system = h2o.copy()
+    h2o_2 = h2o.copy()
+    h2o_2.translate([5.0, 0, 0])
+    system.extend(h2o_2)
+    h2o_3 = h2o.copy()
+    h2o_3.translate([0, 5.0, 0])
+    system.extend(h2o_3)
+    return system
+
+
+# ---------------------------------------------------------------------------
+# Basic structure tests
+# ---------------------------------------------------------------------------
+
+
+def test_analyze_single_molecule(water_molecule):
+    """A single molecule should have one fragment."""
+    dist_mat, bond_mat, frags, _, _ = analyze_structure(water_molecule)
+    assert dist_mat.shape == (3, 3)
+    assert bond_mat.shape == (3, 3)
+    assert len(frags) == 1
+    assert sorted(frags[0]) == [0, 1, 2]
+
+
+def test_analyze_dimer_two_fragments(water_dimer):
+    """Two well-separated molecules should be identified as two fragments."""
+    _, _, frags, centroid_dists, corrected_dists = analyze_structure(water_dimer)
+    assert len(frags) == 2
+    assert len(frags[0]) == 3
+    assert len(frags[1]) == 3
+    # Centroid distance should be approximately 5.0
+    assert centroid_dists[0, 1] > 4.0
+
+
+def test_analyze_three_fragments(three_fragments):
+    """Three well-separated molecules should yield three fragments."""
+    _, _, frags, centroid_dists, corrected_dists = analyze_structure(three_fragments)
+    assert len(frags) == 3
+    for frag in frags:
+        assert len(frag) == 3  # each water has 3 atoms
+
+
+# ---------------------------------------------------------------------------
+# Matrix property tests
+# ---------------------------------------------------------------------------
+
+
+def test_distance_matrix_symmetric(water_molecule):
+    """Distance matrix should be symmetric."""
[…]
