### 2026-02-16 0da42ff
ci: add surfaces and align test steps
Split CI into separate named steps for core, fragments, surfaces,
and align test suites matching pixi environments.
