### 2026-02-17 457f5aa
fix: correct water molecule geometry in tests and remove PDM config
Water fixtures had O-H distances of 1.52 A (exceeding 1.164 A bond
cutoff), causing fragment detection to split each molecule. Use OH2
formula with proper geometry (~0.96 A bonds). Remove leftover PDM
dev-dependencies and scripts sections.
