+++ pyproject.toml
@@ -9,23 +9,6 @@ version-file = "_version.py"
 include = [ "rgpycrumbs/**/*.py", "rgpycrumbs/*.py", "/tests" ]
 
 
-[tool.pdm.dev-dependencies]
-lint = [ "ruff>=0.1.6" ]
-test = [
-  "pytest>=7.4.3",
-  "pytest-cov>=4.1.0",
-  "coverage[toml]>=7.3.2",
-  "pytest-datadir>=1.5.0",
-]
-nbdoc = [ "jupyterlab>=4.0.9", "jupytext>=1.15.2", "ase>=3.22.1" ]
-release_aid = [ "tbump>=6.11.0", "towncrier>=23.11.0" ]
-
-[tool.pdm.scripts]
-check_ruff  = "ruff ."
-ruff_fix    = "ruff --fix --unsafe-fixes ."
-ruff_format = "ruff format ."
-lint        = { composite = [ "ruff_fix", "ruff_format" ] }
-test        = "pytest --cov=rgpycrumbs tests"
 
 [project]
 name = "rgpycrumbs"
