+++ _sources/tools/eon/ptmdisp.rst.txt
@@ -10,7 +10,7 @@ PTM Displacement
 Identifies atoms in a structure file that do not match a specified crystal
 structure (e.g., FCC) using Polyhedral Template Matching (PTM). It outputs the
 0-based indices of these defect atoms (e.g., interstitials), which is useful for
-filtering or setting up atom lists for the displacements used for EON calculations.
+filtering or setting up atom lists for the displacements used for eOn calculations.
 
 .. click:: rgpycrumbs.eon.ptmdisp:main
    :prog: rgpycrumbs eon ptmdisp
