+++ _sources/tools/eon/con_splitter.rst.txt
@@ -22,7 +22,7 @@ To split a trajectory file:
 
 This creates a directory named ``initial_images`` containing ``ipath_000.con``,
 ``ipath_001.con``, etc., along with an ``ipath.dat`` file listing their paths which
-feeds into EON via:
+feeds into eOn via:
 
 .. code:: ini
 
@@ -63,14 +63,14 @@ Centering and Unit Cell Management
 
 The ``--center`` flag translates the center of mass of each image to the origin.
 Use ``--box-diagonal`` to override the simulation cell dimensions (e.g., ``20 20 20``), which ensures all generated images share an identical vacuum box for
-EON calculations.
+eOn calculations.
 
 Initial Path Generation from Generic Files
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-We may also generate valid starting points for an EON NEB calculation from a
+We may also generate valid starting points for an eOn NEB calculation from a
 generic ``.xyz`` file or any ASE compatible input. For instance, for a two-image
-trajectory, to convert a reactant/product pair into the required EON format
+trajectory, to convert a reactant/product pair into the required eOn format
 while fixing coordinate and mapping issues:
 
 .. code:: bash

[…]
