+++ pixi.toml
@@ -10,9 +10,6 @@ platforms = ["linux-64"]
 
 [dependencies]
 python = ">=3.10"
-pip = "23.3.1.*"
-pdm = "2.10.4.*"
-pytest = ">=8.4.2,<9"
 
 [feature.docs.dependencies]
 emacs = ">=30.2,<31"
