### 2026-02-17 60fd193
fix: add rich as core dependency (used in CLI modules)
