+++ pyproject.toml
@@ -30,6 +30,7 @@ classifiers = [
 dependencies = [
   "click>=8.1",
   "numpy>=1.24",
+  "rich>=13.0",
 ]
 dynamic = [ "version" ]
 
