### 2026-02-17 3f88473
build: uv-first workflow, fix conda-only deps and add rich
* build: uv-first workflow, remove conda-only ira from pip extras

- Remove ira_mod from pyproject.toml optional-dependencies (conda-only,
  breaks uv lock/sync). IRA remains pixi-managed.
- Remove pdm/pip from pixi base dependencies (not needed)
- Add test and lint optional-dependency groups to pyproject.toml
- Fix interpolation tests: mark as interpolation (needs scipy), not pure
- Split CI into uv job (pure, interpolation, align on 3.10-3.12) and
  pixi job (fragments, surfaces needing conda packages)

* doc: add development setup a
[…]
