### 2026-02-17 551f8da
doc: release notes for v1.0.0
