### 2026-02-17 a804cd5
ci: add workflow to generate README.md from readme_src.org
Uses the existing scripts/org_to_md.sh to convert orgmode source to
markdown on push to main. Also updates readme_src.org with current
development and release process documentation (uv + hatch-vcs + twine).
