+++ .github/workflows/ci_readme.yml
@@ -1,0 +1,35 @@
+name: Generate README
+
+on:
+  push:
+    branches: [main]
+    paths:
+      - "readme_src.org"
+
+jobs:
+  generate-readme:
+    name: Export readme_src.org to README.md
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Emacs
+        run: |
+          sudo apt-get update -qq
+          sudo apt-get install -y -qq emacs-nox
+
+      - name: Export Org to Markdown
+        run: |
+          bash scripts/org_to_md.sh readme_src.org README.md
+
+      - name: Commit and push if changed
+        run: |
+          git diff --quiet README.md && exit 0
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          git add README.md
+          git commit -m "doc: regenerate README.md from readme_src.org [skip ci]"
+          git push
