### 2026-02-17 d3d8021
doc: regenerate README.md from readme_src.org [skip ci]
