### 2026-02-17 32def44
doc: release notes for v1.0.1
