### 2026-02-17 #275 @HaoZeke
Hi @kysolvik,

Apologies for the delay. I spent some time to think about the package architecture and subsequently refactored both `rgpycrumbs` and the associated plotting library, `chemparseplot` into:

* **`rgpycrumbs`**: Functions as the core computational library and CLI toolkit. It comprises the bulk of the mathematically reviewable content, exposing importable modules for surface fitting (`rgpycrumbs.surfaces`, utilizing JAX-based kernel methods), structural analysis (`rgpycrumbs.geom.analysis`, including fragment detection), IRA structure matching (`rgpycrumbs.geom.ira`), spline interpolation (`rgpycrumbs.interpolation`), and shared data types (`rgpycrumbs.basetypes`).
* *The Launchpad*: The CLI side retains the PEP 723-based architecture, acting as an execution "launchpad" to manage isolated environments for conflicting binaries.

* **`chemparseplot`**: Operates as the downstream parsing and plotting library. It handles outputs from computational chemistry codes (ORCA, eOn) while delegating heavy computation to `rgpycrumbs` to produce publication-quality plots. It now explicitly depends on `rgpycrumbs>=1.0.0`.

Both are now v1.0.1 on PyPI with proper changelogs, uv-first development workflows, and documentation ([rgpycrumbs docs](https://rgpycrumbs.rgoswami.me), [chemparseplot docs](https://chemparseplot.rgoswami.me)).

This separation isolates the objective, algorithmic logic within `rgpycrumbs` suitable for peer review, while delegating subjective parsing and visualization to `chemparseplot`.

Some more specific thoughts:

## Do other researchers use these sub-packages?

To the best of my knowledge, yes, in that these are part of artifacts also on Materials Cloud Archive and seem to be downloaded fairly regularly. A full list is maintained at [rgpycrumbs.rgoswami.me/used_by.html](https://rgpycrumbs.rgoswami.me/used_by.html) but some are repeated here:

  - **[eOn](https://github.com/TheochemUI/eOn)** (TheochemUI) :: the long-timescale molecular dynamics code uses `rgpycrumbs` as its companion diagnostic and visualization suite ([docs](https://eon.theochemui.org/))
  - **[atomistic-cookbook](https://github.com/lab-cosmo/atomistic-cookbook)** (lab-cosmo) :: the community cookbook for atomistic simulations imports `rgpycrumbs.eon.helpers` and `rgpycrumbs.run.jupyter` in the [eOn PET-NEB
  example](https://atomistic-cookbook.org/examples/eon-pet-neb/eon-pet-neb.html)
  - **[metatensor ecosystem-article](https://github.com/metatensor/ecosystem-article)** (metatensor) :: the Metatensor ecosystem paper uses `rgpycrumbs` for eOn NEB plotting and PLUMED FES reconstruction workflows
  - **Several reproduction packages** :: ([gpr_sella_repro](https://github.com/TheochemUI/gpr_sella_repro), [otgpd_repro](https://github.com/TheochemUI/otgpd_repro), [nebmmf_repro](https://github.com/HaoZeke/nebmmf_repro),
  [nebviz_repro](https://github.com/HaoZeke/nebviz_repro), [brms_idrot_repro](https://github.com/HaoZeke/brms_idrot_repro)) — each tied to a specific publication, e.g. the GP-accelerated Sella work (ChemPhysChem, doi:
  [10.1002/cphc.202500730](https://doi.org/10.1002/cphc.202500730))
  - **My doctoral dissertation** (arXiv: [2510.21368](https://arxiv.org/abs/2510.21368), University of Iceland, 2025) uses `rgpycrumbs` throughout

At this stage I would say I'm the sole propagator of the code, but certainly just by adding these packages into learning materials like this [Carpentries style lesson](https://pydissem.rgoswami.me/python_scripts.html) [submitted for review](https://github.com/carpentries-lab/reviews/issues/35) I believe this structure will continue to grow. Also, the atomistic-cookbook and metatensor ecosystem paper are community-facing resources — the cookbook in particular is meant to be a reference for researchers across groups learning to use atomistic simulation tools, and rgpycrumbs is part of that workflow there.

## Does wrapping help or hinder discoverability?

I believe this structure aids discoverability significantly for two reasons:

1. **Shared Conventions**: The modules rely on shared data types. A researcher performing a Nudged Elastic Band (NEB) calculation requires `basetypes.nebpath` to represent the path, `interpolation` for the spline, `surfaces` for the energy landscape, and `geom.analysis` for fragment detection at the saddle point. These tools function as a coherent pipeline for analyzing reaction pathways rather than independent utilities sharing a namespace.
2. **Binary Dependency Management**: The PEP 723 dispatcher addresses a critical packaging issue in computational chemistry. Tools like OVITO (defect analysis) and `tblite` (electronic structure) possess conflicting binary requirements. Separating them into distinct packages would fail to resolve the conflict; the unified dispatcher architecture provides the necessary solution.

The refactoring into `rgpycrumbs` (computation) and `chemparseplot` (parsing/plotting) directly addresses the concern regarding "loosely connected" scripts. The library modules in `rgpycrumbs` now exhibit clear cohesion as they all serve the computational analysis of reaction pathways and molecular structures.
