### 2026-02-18 #275 @kysolvik
Hi again, @HaoZeke! Thanks for your patience, and I really appreciate the time you've taken to answer my questions. I think you make a good case for the combined functionality of `rgpycrumbs` and show how it can support scientific workflows. With the refactor and some of the usage examples you've highlighted, we'd be happy to review `rgpycrumbs`.

I've tagged this as submission-requested. Be sure to reference this pre-submission issue in the full submission.

Once you submit, we'll complete the [full pre-review checks](https://www.pyopensci.org/software-peer-review/how-to/editor-in-chief-guide.html#editor-checklist-template). Note that we've recently adopted policy for use of generative AI/LLMs, which requires disclosure of if and how generative AI was used to produce material in the submission. You can read more about that policy and view the generative AI submission checklist [here in our Peer Review Guide](https://www.pyopensci.org/software-peer-review/our-process/policies.html#policy-for-use-of-generative-ai-llms).

Let me know if you have any questions, and we'll look out for your full submission!
