### 2026-02-19 df9b63b
feat(gp): add in the uncertainty
