### 2026-02-19 bb608e8
feat(gp): add in the uncertainty
