+++ rgpycrumbs/surfaces.py
@@ -80,7 +80,8 @@ def _tps_solve(x, y, sm):
     rhs = jnp.concatenate([y, jnp.zeros(M, dtype=jnp.float32)])
 
     coeffs = jnp.linalg.solve(lhs, rhs)
-    return coeffs[:N], coeffs[N:]
+    lhs_inv = jnp.linalg.inv(lhs)
+    return coeffs[:N], coeffs[N:], lhs_inv
 
 
 @jit
@@ -93,6 +94,19 @@ def _tps_predict(x_query, x_obs, w, v):
         [jnp.ones((x_query.shape[0], 1), dtype=jnp.float32), x_query], axis=1
     )
     return K_q @ w + P_q @ v
+
+
+@jit
+def _tps_var(x_query, x_obs, lhs_inv):
+    d2 = jnp.sum((x_query[:, None, :] - x_obs[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K_q = r**2 * jnp.log(r)
+    P_q = jnp.concatenate(
+        [jnp.ones((x_query.shape[0], 1), dtype=jnp.float32), x_query], axis=1
+    )
+    KP_q = jnp.concatenate([K_q, P_q], axis=1)
+    var = -jnp.sum((KP_q @ lhs_inv) * KP_q, axis=1)
+    return jnp.maximum(var, 0.0)
 
 
 class FastTPS:
@@ -121,7 +135,7 @@ class FastTPS:
         else:
             self.sm = init_sm
 
-        self.w, self.v = _tps_solve(self.x_obs, self.y_obs, self.sm)
+        self.w, self.v, self.K_inv = _tps_solve(self.x_obs, self.y_obs, self.sm)
 
     def __call__(self, x_query, chunk_size=500):
         """
@@ -133,6 +147,14 @@ class FastTPS:
             chunk = x_query[i : i + chunk_size]
             preds.append(_tps_predict(chunk, self.x_obs, self.w, self.v))
         return jnp.concatenate(preds, axis=0)
+
+    def predict_var(self, x_query, chunk_size=500):
+        x_query = jnp.asarray(x_query, dtype=jnp.float32)
+        vars_list = []
+        for i in range(0, x_query.shape[0], chunk_size):
+            chunk = x_query[i : i + chunk_size]
+            vars_list.append(_tps_var(chunk, self.x_obs, self.K_inv))
+        return jnp.concatenate(vars_list, axis=0)
 
 
 # ==============================================================================
@@ -169,7 +191,11 @@ def _matern_solve(x, y, sm, length_scale):
     # mean(y) before fitting and add it back after.
     L = jnp.linalg.cholesky(K)
     alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
-    return alpha
+    
+    eye = jnp.eye(K.shape[0])
+    L_inv = jnp.linalg.solve(L, eye)
+    K_inv = L_inv.T @ L_inv
+    return alpha, K_inv
 
 
 @jit
@@ -183,6 +209,16 @@ def _matern_predict(x_query, x_obs, alpha, length_scale):
     )
 
     return K_q @ alpha
+
+
+@jit
+def _matern_var(x_query, x_obs, K_inv, length_scale):
+    d2 = jnp.sum((x_query[:, None, :] - x_obs[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    sqrt5_r_l = jnp.sqrt(5.0) * r / length_scale
[…]
