### 2026-02-19 570b741
feat(path): project orthogonal to RMSD(R)
