### 2026-02-19 ab4a78f
chore(path): expand grid to cover extra points
also change the default uncertainity and legend positioning
