### 2026-02-19 66f7d2c
chore(pltneb): set according to projection
