### 2026-02-20 03e4407
chore(pltneb): lint and remove debugging
