### 2026-02-20 d80e533
feat(dev): add details on working with local
packages
