### 2026-02-20 e565424
feat(verbose): add a flag
