### 2026-02-20 e4793ab
feat(verbose): add a flag
