+++ .taplo.toml
@@ -1,19 +1,19 @@
 [formatting]
-align_entries = true
-align_comments = true
-array_trailing_comma = true
-array_auto_expand = true
-array_auto_collapse = true
-compact_arrays = false
+align_comments        = true
+align_entries         = true
+allowed_blank_lines   = 2
+array_auto_collapse   = true
+array_auto_expand     = true
+array_trailing_comma  = true
+column_width          = 80
+compact_arrays        = false
+compact_entries       = false
 compact_inline_tables = false
-inline_table_expand = true
-compact_entries = false
-column_width = 80
-indent_tables = false
-indent_entries = false
-indent_string = "  "
-trailing_newline = true
-reorder_keys = false
-reorder_arrays = false
-allowed_blank_lines = 2
-crlf = false
+crlf                  = false
+indent_entries        = false
+indent_string         = "  "
+indent_tables         = false
+inline_table_expand   = true
+reorder_arrays        = false
+reorder_keys          = true
+trailing_newline      = true
