### 2026-02-20 6baa092
chore(py): apply lint consistently
