+++ rgpycrumbs/geom/ira.py
@@ -2,10 +2,10 @@
 # PYTHONPATH export.. after that's cleared up this will probably be a dependency
 # [1]: https://github.com/mammasmias/IterativeRotationsAssignments
 import dataclasses
+from collections import Counter
 
 import ira_mod
 import numpy as np
-from collections import Counter
 
 
 @dataclasses.dataclass
@@ -62,9 +62,7 @@ def calculate_rmsd(atm1, atm2, k_factor=2.8):
     """Calculates RMSD using do_ira."""
     ira_comp = do_ira(atm1, atm2, k_factor)
     atm2_coords_permuted = atm2.get_positions()[ira_comp.perm]
-    atm2_coords_transformed = (
-        np.dot(atm2_coords_permuted, ira_comp.rot) + ira_comp.trans
-    )
+    atm2_coords_transformed = np.dot(atm2_coords_permuted, ira_comp.rot) + ira_comp.trans
     n_atoms = len(atm1.get_positions())
     squared_distances = np.sum(
         (atm1.get_positions() - atm2_coords_transformed) ** 2, axis=1
