### 2026-02-20 bfa3562
chore(py): apply lint consistently
