+++ rgpycrumbs/surfaces.py
@@ -1,3 +1,5 @@
+import logging
+
 import jax
 import jax.numpy as jnp
 import jax.scipy.optimize as jopt
@@ -23,7 +25,8 @@ def safe_cholesky_solve(K, y, noise_scalar, jitter_steps=3):
             alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
             log_det = 2.0 * jnp.sum(jnp.log(jnp.diag(L)))
             return alpha, log_det
-        except:
+        except Exception as e:
+            logging.debug(f"Cholesky failed: {e}")
             continue
 
     # Fallback for compilation safety (NaN propagation)
@@ -191,7 +194,7 @@ def _matern_solve(x, y, sm, length_scale):
     # mean(y) before fitting and add it back after.
     L = jnp.linalg.cholesky(K)
     alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
-    
+
     eye = jnp.eye(K.shape[0])
     L_inv = jnp.linalg.solve(L, eye)
     K_inv = L_inv.T @ L_inv
@@ -204,9 +207,7 @@ def _matern_predict(x_query, x_obs, alpha, length_scale):
     r = jnp.sqrt(d2 + 1e-12)
 
     sqrt5_r_l = jnp.sqrt(5.0) * r / length_scale
-    K_q = (1.0 + sqrt5_r_l + (5.0 * r**2) / (3.0 * length_scale**2)) * jnp.exp(
-        -sqrt5_r_l
-    )
+    K_q = (1.0 + sqrt5_r_l + (5.0 * r**2) / (3.0 * length_scale**2)) * jnp.exp(-sqrt5_r_l)
 
     return K_q @ alpha
 
@@ -259,7 +260,9 @@ class FastMatern:
             self.ls = init_ls
             self.noise = init_noise
 
-        self.alpha, self.K_inv = _matern_solve(self.x_obs, y_centered, self.noise, self.ls)
+        self.alpha, self.K_inv = _matern_solve(
+            self.x_obs, y_centered, self.noise, self.ls
+        )
 
     def __call__(self, x_query, chunk_size=500):
         """
@@ -413,7 +416,9 @@ class GradientMatern:
         else:
             self.ls, self.noise = init_ls, init_noise
 
-        self.alpha, self.K_inv = _grad_matern_solve(self.x, self.y_full, self.noise, self.ls)
+        self.alpha, self.K_inv = _grad_matern_solve(
+            self.x, self.y_full, self.noise, self.ls
+        )
 
     def __call__(self, x_query, chunk_size=500):
         x_query = jnp.asarray(x_query, dtype=jnp.float32)
@@ -457,7 +462,7 @@ def _imq_solve(x, y, sm, epsilon):
     K = K + jnp.eye(x.shape[0]) * sm
     L = jnp.linalg.cholesky(K)
     alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
-    
+
     eye = jnp.eye(K.shape[0])
     L_inv = jnp.linalg.solve(L, eye)
     K_inv = L_inv.T @ L_inv
@@ -509,7 +514,9 @@ class FastIMQ:
         else:
             self.epsilon, self.noise = init_eps, init_noise
 
-        self.alpha, self.K_inv = _imq_solve(self.x_obs, y_centered, self.noise, self.epsilon)
+        self.alpha, self.K_inv = _imq_solve(
+            self.x_obs, y_centered, self.noise, self.epsilon
+        )
 
     def __call__(self, x_query, chunk_size=500):
         x_query = jnp.asarray(x_query, dtype=jnp.float32)
[…]
