+++ rgpycrumbs/basetypes.py
@@ -1,6 +1,6 @@
 import datetime
 from collections import namedtuple
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 
 import numpy as np
 
@@ -72,7 +72,7 @@ class MolGeom:
 class SaddleMeasure:
     pes_calls: int = 0
     iter_steps: int = 0
-    tot_time: datetime.timedelta = datetime.timedelta(0).total_seconds()
+    tot_time: float = field(default_factory=lambda: datetime.timedelta(0).total_seconds())
     saddle_energy: float = np.nan
     saddle_fmax: float = np.nan
     success: bool = False
