### 2026-02-20 b0b4c81
chore(py): finish removing cruft
