### 2026-02-20 7984dc5
chore(viz): default to 50% var
