### 2026-02-20 5ba403e
feat(surfaces): add a nystrom kernel
Way faster, somewhat less accurate
