+++ rgpycrumbs/surfaces.py
@@ -2,7 +2,9 @@ import logging
 
 import jax
 import jax.numpy as jnp
+import jax.scipy.linalg as jlinalg
 import jax.scipy.optimize as jopt
+import numpy as np
 from jax import jit, vmap
 
 # Force float32 for speed/viz
@@ -1056,6 +1058,177 @@ class GradientRQ:
         return jnp.concatenate(vars_list, axis=0)
 
 
+# ==============================================================================
+# NYSTRÖM GRADIENT-ENHANCED IMQ
+# ==============================================================================
+
+
+@jit
+def _stable_nystrom_grad_imq_solve(x, y_full, x_inducing, noise_scalar, epsilon):
+    N = x.shape[0]
+    M = x_inducing.shape[0]
+    D_plus_1 = x.shape[1] + 1
+
+    # 1. Base inducing point matrix (K_mm)
+    K_mm_blocks = k_matrix_imq_grad_map(x_inducing, x_inducing, epsilon)
+    K_mm = K_mm_blocks.transpose(0, 2, 1, 3).reshape(M * D_plus_1, M * D_plus_1)
+
+    # Heavier jitter is required for float32 gradient matrices
+    jitter = (noise_scalar + 1e-4) * jnp.eye(M * D_plus_1)
+    K_mm = K_mm + jitter
+
+    # Stable Cholesky factorization of K_mm
+    L = jnp.linalg.cholesky(K_mm)
+
+    # 2. Cross-covariance matrix (K_mn)
+    K_nm_blocks = k_matrix_imq_grad_map(x, x_inducing, epsilon)
+    K_nm = K_nm_blocks.transpose(0, 2, 1, 3).reshape(N * D_plus_1, M * D_plus_1)
+    K_mn = K_nm.T
+
+    # 3. Normalized features: V = L^{-1} K_mn
+    # This avoids squaring the condition number!
+    V = jlinalg.solve_triangular(L, K_mn, lower=True)
+
+    # 4. Inner system: S = V V^T + sigma^2 I
+    sigma2 = noise_scalar + 1e-6
+    S = V @ V.T + sigma2 * jnp.eye(M * D_plus_1)
+    L_S = jnp.linalg.cholesky(S)
+
+    # 5. Solve for weights: alpha_m = L^{-T} S^{-1} V y
+    Vy = V @ y_full.flatten()
+    beta = jlinalg.cho_solve((L_S, True), Vy)
+    alpha_m = jlinalg.solve_triangular(L.T, beta, lower=False)
+
+    # 6. Effective precision matrix for predictive variance
+    # W = L^{-T} (I - sigma^2 S^{-1}) L^{-1}
+    I_M = jnp.eye(M * D_plus_1)
+    S_inv = jlinalg.cho_solve((L_S, True), I_M)
+    inner = I_M - sigma2 * S_inv
+
+    L_inv = jlinalg.solve_triangular(L, I_M, lower=True)
+    W = L_inv.T @ inner @ L_inv
+
+    return alpha_m, W
+
+
+@jit
+def _nystrom_grad_imq_predict(x_query, x_inducing, alpha_m, epsilon):
+    def get_query_row(xq, xo):
+        kee = imq_kernel_elem(xq, xo, epsilon)
+        ked = jax.grad(imq_kernel_elem, argnums=1)(xq, xo, epsilon)
+        return jnp.concatenate([kee[None], ked])
+
+    K_qm = vmap(vmap(get_query_row, (None, 0)), (0, None))(x_query, x_inducing)
+    Q, M, D_plus_1 = K_qm.shape
+    return K_qm.reshape(Q, M * D_plus_1) @ alpha_m
+
+
+@jit
[…]
