### 2026-02-20 25f500c
test(surfaces): add comprehensive tests for gradient models and optimization
