+++ tests/test_surfaces.py
@@ -14,6 +14,7 @@ from rgpycrumbs.surfaces import (  # noqa: E402
     GradientMatern,
     GradientRQ,
     GradientSE,
+    NystromGradientIMQ,
     _imq_kernel_matrix,
     _matern_kernel_matrix,
     _tps_kernel_matrix,
@@ -70,6 +71,7 @@ def test_get_surface_model_returns_classes():
     assert get_surface_model("grad_rq") is GradientRQ
     assert get_surface_model("grad_se") is GradientSE
     assert get_surface_model("grad_imq") is GradientIMQ
+    assert get_surface_model("grad_imq_ny") is NystromGradientIMQ
 
 
 def test_get_surface_model_default():
@@ -218,6 +220,73 @@ def test_fast_imq_interpolates_training_points(simple_2d_data):
 
 
 # ---------------------------------------------------------------------------
+# Model instantiation and prediction tests (gradient-enhanced models)
+# ---------------------------------------------------------------------------
+
+
+def test_gradient_matern_fit_and_predict(gradient_2d_data, query_points):
+    """GradientMatern should fit and produce finite predictions."""
+    x, y, grads = gradient_2d_data
+    model = GradientMatern(x, y, gradients=grads, optimize=False, length_scale=1.0)
+    preds = model(query_points)
+    assert preds.shape == (3,)
+    assert jnp.all(jnp.isfinite(preds))
+
+
+def test_gradient_imq_fit_and_predict(gradient_2d_data, query_points):
+    """GradientIMQ should fit and produce finite predictions."""
+    x, y, grads = gradient_2d_data
+    model = GradientIMQ(x, y, gradients=grads, optimize=False, length_scale=1.0)
+    preds = model(query_points)
+    assert preds.shape == (3,)
+    assert jnp.all(jnp.isfinite(preds))
+
+
+def test_gradient_se_fit_and_predict(gradient_2d_data, query_points):
+    """GradientSE should fit and produce finite predictions."""
+    x, y, grads = gradient_2d_data
+    model = GradientSE(x, y, gradients=grads, optimize=False, length_scale=1.0)
+    preds = model(query_points)
+    assert preds.shape == (3,)
+    assert jnp.all(jnp.isfinite(preds))
+
+
+def test_gradient_rq_fit_and_predict(gradient_2d_data, query_points):
+    """GradientRQ should fit and produce finite predictions."""
+    x, y, grads = gradient_2d_data
+    model = GradientRQ(x, y, gradients=grads, optimize=False, length_scale=1.0)
+    preds = model(query_points)
+    assert preds.shape == (3,)
+    assert jnp.all(jnp.isfinite(preds))
+
+
+def test_nystrom_gradient_imq_fit_and_predict(gradient_2d_data, query_points):
+    """NystromGradientIMQ should fit and produce finite predictions."""
+    x, y, grads = gradient_2d_data
+    # Use n_inducing >= N_total to test basic fit
+    model = NystromGradientIMQ(
+        x, y, gradients=grads, n_inducing=10, optimize=False, length_scale=1.0
+    )
+    preds = model(query_points)
+    assert preds.shape == (3,)
+    assert jnp.all(jnp.isfinite(preds))
+
+
+def test_nystrom_gradient_imq_path_sampling(gradient_2d_data):
+    """Test the path-based sampling logic in NystromGradientIMQ."""
+    x, y, grads = gradient_2d_data
+    # Create synthetic "paths"
+    nimags = 2
+    # Simple data has 5 points, so 2 paths of 2 and 1 residual
+    model = NystromGradientIMQ(
+        x, y, gradients=grads, n_inducing=2, nimags=nimags, optimize=False
+    )
+    # The sampling logic should produce an x_inducing
[…]
