### 2026-02-20 0447cae
refactor(surfaces): introduce base classes for standard and gradient models
