+++ rgpycrumbs/surfaces.py
@@ -16,7 +16,19 @@ jax.config.update("jax_enable_x64", False)
 
 
 def safe_cholesky_solve(K, y, noise_scalar, jitter_steps=3):
-    """Retries Cholesky with increasing jitter if it fails."""
+    """
+    Retries Cholesky decomposition with increasing jitter if it fails.
+
+    Args:
+        K: Covariance matrix.
+        y: Observation vector.
+        noise_scalar: Initial noise level.
+        jitter_steps: Number of retry attempts with increasing jitter.
+
+    Returns:
+        tuple: (alpha, log_det) where alpha is the solution vector and
+               log_det is the log determinant of the jittered matrix.
+    """
     N = K.shape[0]
 
     # Try successively larger jitters: 1e-6, 1e-5, 1e-4
@@ -36,6 +48,17 @@ def safe_cholesky_solve(K, y, noise_scalar, jitter_steps=3):
 
 
 def generic_negative_mll(K, y, noise_scalar):
+    """
+    Calculates the negative Marginal Log-Likelihood (MLL).
+
+    Args:
+        K: Covariance matrix.
+        y: Observation vector.
+        noise_scalar: Noise level for regularization.
+
+    Returns:
+        float: The negative MLL value, or a high penalty if Cholesky fails.
+    """
     alpha, log_det = safe_cholesky_solve(K, y, noise_scalar)
 
     data_fit = 0.5 * jnp.dot(y.flatten(), alpha.flatten())
@@ -44,6 +67,194 @@ def generic_negative_mll(K, y, noise_scalar):
     cost = data_fit + complexity
     # heavy penalty if Cholesky failed (NaN)
     return jnp.where(jnp.isnan(cost), 1e9, cost)
+
+
+# ==============================================================================
+# BASE CLASSES
+# ==============================================================================
+
+
+class BaseSurface:
+    """
+    Abstract base class for standard (non-gradient) surface models.
+
+    Derived classes must implement `_fit`, `_solve`, `_predict_chunk`, and `_var_chunk`.
+    """
+
+    def __init__(
+        self, x_obs, y_obs, smoothing=1e-3, length_scale=None, optimize=True, **kwargs
+    ):
+        """
+        Initializes and fits the surface model.
+
+        Args:
+            x_obs: Training inputs (N, D).
+            y_obs: Training observations (N,).
+            smoothing: Initial noise/smoothing parameter.
+            length_scale: Initial length scale parameter(s).
+            optimize: Whether to optimize parameters via MLE.
+            **kwargs: Additional model-specific parameters.
+        """
+        self.x_obs = jnp.asarray(x_obs, dtype=jnp.float32)
+        self.y_obs = jnp.asarray(y_obs, dtype=jnp.float32)
+        # Center the data
+        self.y_mean = jnp.mean(self.y_obs)
+        self.y_centered = self.y_obs - self.y_mean
+
+        self._fit(smoothing, length_scale, optimize)
+        self._solve()
+
+    def _fit(self, smoothing, length_scale, optimize):
+        """Internal method to perform parameter optimization."""
+        raise NotImplementedError
+
+    def _solve(self):
+        """Internal method to solve the linear system for weights."""
+        raise NotImplementedError
+
+    def __call__(self, x_query, chunk_size=500):
+        """
+        Predict values at query points.
+
+        Args:
+            x_query: Query inputs (M, D).
+            chunk_size: Number of points to process per batch to avoid OOM.
+
+        Returns:
[…]
