### 2026-02-20 1cdccb9
refactor(surfaces): split surfaces.py into a structured package
