+++ rgpycrumbs/surfaces/standard.py
@@ -1,0 +1,308 @@
+import jax.numpy as jnp
+import jax.scipy.optimize as jopt
+from jax import jit
+
+from rgpycrumbs.surfaces._base import BaseSurface, generic_negative_mll
+from rgpycrumbs.surfaces._kernels import (
+    _imq_kernel_matrix,
+    _matern_kernel_matrix,
+    _tps_kernel_matrix,
+)
+
+# ==============================================================================
+# TPS HELPERS
+# ==============================================================================
+
+
+def negative_mll_tps(log_params, x, y):
+    # TPS only really has a smoothing parameter to tune in this context
+    # (Length scale is inherent to the radial basis).
+    smoothing = jnp.exp(log_params[0])
+    K = _tps_kernel_matrix(x)
+    return generic_negative_mll(K, y, smoothing)
+
+
+@jit
+def _tps_solve(x, y, sm):
+    K = _tps_kernel_matrix(x)
+    K = K + jnp.eye(x.shape[0]) * sm
+
+    # Polynomial Matrix
+    N = x.shape[0]
+    P = jnp.concatenate([jnp.ones((N, 1), dtype=jnp.float32), x], axis=1)
+    M = P.shape[1]
+
+    # Solve System
+    zeros = jnp.zeros((M, M), dtype=jnp.float32)
+    top = jnp.concatenate([K, P], axis=1)
+    bot = jnp.concatenate([P.T, zeros], axis=1)
+    lhs = jnp.concatenate([top, bot], axis=0)
+    rhs = jnp.concatenate([y, jnp.zeros(M, dtype=jnp.float32)])
+
+    coeffs = jnp.linalg.solve(lhs, rhs)
+    lhs_inv = jnp.linalg.inv(lhs)
+    return coeffs[:N], coeffs[N:], lhs_inv
+
+
+@jit
+def _tps_predict(x_query, x_obs, w, v):
+    d2 = jnp.sum((x_query[:, None, :] - x_obs[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K_q = r**2 * jnp.log(r)
+
+    P_q = jnp.concatenate(
+        [jnp.ones((x_query.shape[0], 1), dtype=jnp.float32), x_query], axis=1
+    )
+    return K_q @ w + P_q @ v
+
+
+@jit
+def _tps_var(x_query, x_obs, lhs_inv):
+    d2 = jnp.sum((x_query[:, None, :] - x_obs[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K_q = r**2 * jnp.log(r)
+    P_q = jnp.concatenate(
+        [jnp.ones((x_query.shape[0], 1), dtype=jnp.float32), x_query], axis=1
+    )
+    KP_q = jnp.concatenate([K_q, P_q], axis=1)
+    var = -jnp.sum((KP_q @ lhs_inv) * KP_q, axis=1)
+    return jnp.maximum(var, 0.0)
+
+
+class FastTPS:
+    """
+    Thin Plate Spline (TPS) surface implementation.
+    Includes a polynomial mean function and supports smoothing optimization.
+    """
+
+    def __init__(self, x_obs, y_obs, smoothing=1e-3, optimize=True, **_kwargs):
+        """
+        Initializes the TPS model.
+
+        Args:
+            x_obs: Training inputs (N, D).
+            y_obs: Training observations (N,).
+            smoothing: Initial smoothing parameter.
+            optimize: Whether to optimize the smoothing parameter.
+        """
+        self.x_obs = jnp.asarray(x_obs, dtype=jnp.float32)
+        self.y_obs = jnp.asarray(y_obs, dtype=jnp.float32)
+
[…]
