### 2026-02-20 8520369
refactor(surfaces): split surfaces.py into a structured package
