+++ rgpycrumbs/surfaces/_base.py
@@ -1,0 +1,240 @@
+import logging
+
+import jax.numpy as jnp
+
+
+def safe_cholesky_solve(K, y, noise_scalar, jitter_steps=3):
+    """
+    Retries Cholesky decomposition with increasing jitter if it fails.
+
+    Args:
+        K: Covariance matrix.
+        y: Observation vector.
+        noise_scalar: Initial noise level.
+        jitter_steps: Number of retry attempts with increasing jitter.
+
+    Returns:
+        tuple: (alpha, log_det) where alpha is the solution vector and
+               log_det is the log determinant of the jittered matrix.
+    """
+    N = K.shape[0]
+
+    # Try successively larger jitters: 1e-6, 1e-5, 1e-4
+    for i in range(jitter_steps):
+        jitter = (noise_scalar + 10 ** (-6 + i)) * jnp.eye(N)
+        try:
+            L = jnp.linalg.cholesky(K + jitter)
+            alpha = jnp.linalg.solve(L.T, jnp.linalg.solve(L, y))
+            log_det = 2.0 * jnp.sum(jnp.log(jnp.diag(L)))
+            return alpha, log_det
+        except Exception as e:
+            logging.debug(f"Cholesky failed: {e}")
+            continue
+
+    # Fallback for compilation safety (NaN propagation)
+    return jnp.zeros_like(y), jnp.nan
+
+
+def generic_negative_mll(K, y, noise_scalar):
+    """
+    Calculates the negative Marginal Log-Likelihood (MLL).
+
+    Args:
+        K: Covariance matrix.
+        y: Observation vector.
+        noise_scalar: Noise level for regularization.
+
+    Returns:
+        float: The negative MLL value, or a high penalty if Cholesky fails.
+    """
+    alpha, log_det = safe_cholesky_solve(K, y, noise_scalar)
+
+    data_fit = 0.5 * jnp.dot(y.flatten(), alpha.flatten())
+    complexity = 0.5 * log_det
+
+    cost = data_fit + complexity
+    # heavy penalty if Cholesky failed (NaN)
+    return jnp.where(jnp.isnan(cost), 1e9, cost)
+
+
+class BaseSurface:
+    """
+    Abstract base class for standard (non-gradient) surface models.
+
+    Derived classes must implement `_fit`, `_solve`, `_predict_chunk`, and `_var_chunk`.
+    """
+
+    def __init__(
+        self, x_obs, y_obs, smoothing=1e-3, length_scale=None, optimize=True, **_kwargs
+    ):
+        """
+        Initializes and fits the surface model.
+
+        Args:
+            x_obs: Training inputs (N, D).
+            y_obs: Training observations (N,).
+            smoothing: Initial noise/smoothing parameter.
+            length_scale: Initial length scale parameter(s).
+            optimize: Whether to optimize parameters via MLE.
+            **kwargs: Additional model-specific parameters.
+        """
+        self.x_obs = jnp.asarray(x_obs, dtype=jnp.float32)
+        self.y_obs = jnp.asarray(y_obs, dtype=jnp.float32)
+        # Center the data
+        self.y_mean = jnp.mean(self.y_obs)
+        self.y_centered = self.y_obs - self.y_mean
+
+        self._fit(smoothing, length_scale, optimize)
+        self._solve()
+
+    def _fit(self, smoothing, length_scale, optimize):
+        """Internal method to perform parameter optimization."""
+        raise NotImplementedError
+
+    def _solve(self):
+        """Internal method to solve the linear system for weights."""
+        raise NotImplementedError
+
+    def __call__(self, x_query, chunk_size=500):
+        """
+        Predict values at query points.
+
+        Args:
+            x_query: Query inputs (M, D).
[…]
