+++ rgpycrumbs/surfaces/_kernels.py
@@ -1,0 +1,150 @@
+import jax
+import jax.numpy as jnp
+from jax import jit, vmap
+
+# ==============================================================================
+# TPS KERNELS
+# ==============================================================================
+
+
+@jit
+def _tps_kernel_matrix(x):
+    d2 = jnp.sum((x[:, None, :] - x[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+    K = r**2 * jnp.log(r)
+    return K
+
+
+# ==============================================================================
+# MATERN KERNELS
+# ==============================================================================
+
+
+@jit
+def _matern_kernel_matrix(x, length_scale):
+    d2 = jnp.sum((x[:, None, :] - x[None, :, :]) ** 2, axis=-1)
+    r = jnp.sqrt(d2 + 1e-12)
+
+    # Matérn 5/2 Kernel
+    # k(r) = (1 + sqrt(5)r/l + 5r^2/3l^2) * exp(-sqrt(5)r/l)
+    sqrt5_r_l = jnp.sqrt(5.0) * r / length_scale
+    K = (1.0 + sqrt5_r_l + (5.0 * r**2) / (3.0 * length_scale**2)) * jnp.exp(-sqrt5_r_l)
+    return K
+
+
+def matern_kernel_elem(x1, x2, length_scale=1.0):
+    d2 = jnp.sum((x1 - x2) ** 2)
+    r = jnp.sqrt(d2 + 1e-12)
+    ls = jnp.squeeze(length_scale)
+    sqrt5_r_l = jnp.sqrt(5.0) * r / ls
+    val = (1.0 + sqrt5_r_l + (5.0 * r**2) / (3.0 * ls**2)) * jnp.exp(-sqrt5_r_l)
+    return val
+
+
+def full_covariance_matern(x1, x2, length_scale):
+    k_ee = matern_kernel_elem(x1, x2, length_scale)
+    k_ed = jax.grad(matern_kernel_elem, argnums=1)(x1, x2, length_scale)
+    k_de = jax.grad(matern_kernel_elem, argnums=0)(x1, x2, length_scale)
+    k_dd = jax.jacfwd(jax.grad(matern_kernel_elem, argnums=1), argnums=0)(
+        x1, x2, length_scale
+    )
+    row1 = jnp.concatenate([k_ee[None], k_ed])
+    row2 = jnp.concatenate([k_de[:, None], k_dd], axis=1)
+    return jnp.concatenate([row1[None, :], row2], axis=0)
+
+
+k_matrix_matern_grad_map = vmap(
+    vmap(full_covariance_matern, (None, 0, None)), (0, None, None)
+)
+
+
+# ==============================================================================
+# IMQ KERNELS
+# ==============================================================================
+
+
+@jit
+def _imq_kernel_matrix(x, epsilon):
+    d2 = jnp.sum((x[:, None, :] - x[None, :, :]) ** 2, axis=-1)
+    K = 1.0 / jnp.sqrt(d2 + epsilon**2)
+    return K
+
+
+def imq_kernel_elem(x1, x2, epsilon=1.0):
+    d2 = jnp.sum((x1 - x2) ** 2)
+    val = 1.0 / jnp.sqrt(d2 + epsilon**2)
+    return val
+
+
+def full_covariance_imq(x1, x2, epsilon):
+    k_ee = imq_kernel_elem(x1, x2, epsilon)
+    k_ed = jax.grad(imq_kernel_elem, argnums=1)(x1, x2, epsilon)
+    k_de = jax.grad(imq_kernel_elem, argnums=0)(x1, x2, epsilon)
+    k_dd = jax.jacfwd(jax.grad(imq_kernel_elem, argnums=1), argnums=0)(x1, x2, epsilon)
[…]
