### 2026-02-20 618f9d3
docs(surfaces): update changelog and fix rendering issues
