+++ .github/workflows/ci_test.yml
@@ -19,6 +19,8 @@ jobs:
       - uses: astral-sh/setup-uv@v5
         with:
           python-version: ${{ matrix.python }}
+      - name: Install dependencies
+        run: uv sync --all-extras
       - name: Run pure tests
         run: uv run --extra test pytest -m pure
       - name: Run interpolation tests
