+++ pyproject.toml
@@ -35,7 +35,7 @@ interpolation = [ "scipy>=1.11" ]
 surfaces      = [ "jax>=0.4" ]
 # ira_mod is conda-only (via pixi); not a pip-installable extra
 lint = [ "ruff>=0.1.6" ]
-test = [ "pytest>=9.0", "pytest-cov>=4.1.0" ]
+test = [ "pytest>=9.0", "pytest-cov>=4.1.0", "pychum", "chemparseplot" ]
 
 [project.scripts]
 rgpycrumbs = "rgpycrumbs.cli:main"
