### 2026-02-21 6c84278
rgpycrumbs: Prepare v1.1.0 release and synchronize dispatcher
This commit consolidates the transition to v1.1.0. The primary changes
involve fixing a signature mismatch in 'analyze_structure' within
'rgpycrumbs.geom.analysis', which now correctly returns a 5-tuple
including centroid and corrected distances. All associated tests in
'tests/test_analysis.py' have been updated.

Furthermore, the NWChem input generator has been refactored into a
dispatcher command that delegates core logic to the 'pychum' library.
A new ORCA input generator shim has also been added to the dispatcher
architecture.


[…]
