+++ rgpycrumbs/eon/generate_nwchem_input.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
-
 # /// script
 # requires-python = ">=3.11"
 # dependencies = [
 #   "ase>=3.25",
 #   "click>=8.2.1",
 #   "rich",
+#   "pychum",
 # ]
 # ///
 
@@ -14,20 +14,9 @@ import logging
 import sys
 from pathlib import Path
 
-try:
-    import click
-    from ase.io import read as ase_read
-    from rich.logging import RichHandler
-except ImportError:
-    print(
-        "Error: Required libraries (ase, click, rich) are not installed.",
-        file=sys.stderr,
-    )
-    print(
-        "Please run this script using a PEP 723-compliant runner like 'uv run <script_name>.py'",
-        file=sys.stderr,
-    )
-    sys.exit(1)
+import click
+from rich.logging import RichHandler
+from pychum import render_nwchem
 
 # --- Logging Configuration ---
 logging.basicConfig(
@@ -36,80 +25,6 @@ logging.basicConfig(
     datefmt="[%X]",
     handlers=[RichHandler(rich_tracebacks=True, markup=True)],
 )
-
-
-def generate_nwchem_input(
-    pos_path: Path,
-    settings_path: Path,
-    socket_address: str,
-    *,
-    unix_mode: bool,
-    mem_in_gb: int,
-    output_path: Path,
-):
-    """
-    Generates a complete NWChem input file by combining atom types from a .con
-    file with a user-provided settings file.
-
-    Args:
-        pos_path: Path to the input geometry file (e.g., pos.con).
-        settings_path: Path to the user's NWChem settings file.
-        socket_address: The address for the socket (path for UNIX, host:port for TCP).
-        unix_mode: Boolean indicating if a UNIX socket should be used.
-        mem_in_gb: Memory to allocate for the NWChem calculation in Gigabytes.
-        output_path: The path for the generated NWChem input file.
-    """
-    logging.info(f"Reading atom types from: [cyan]{pos_path}[/cyan]")
-    try:
-        atoms = ase_read(pos_path)
-        if len(atoms) == 0:
-            logging.critical(f"Input geometry file '{pos_path}' contains no atoms.")
-            sys.exit(1)
-    except FileNotFoundError:
-        logging.critical(f"Geometry file not found at '{pos_path}'")
-        sys.exit(1)
-    except Exception as e:
-        logging.critical(f"Failed to read geometry file with ASE: {e}")
-        sys.exit(1)
-
-    if not settings_path.is_file():
-        logging.critical(f"NWChem settings file not found at '{settings_path}'")
-        sys.exit(1)
-
-    logging.info(f"Generating NWChem input file: [cyan]{output_path}[/cyan]")
-    try:
-        with open(output_path, "w") as f:
-            f.write("start nwchem_socket_job\n")
-            f.write('title "NWChem Server for eOn"\n\n')
-
-            f.write(f"memory {mem_in_gb} gb\n\n")
-
-            # This geometry block is only a template for memory allocation.  The
-            # atom types and count are what matter. Rather than confuse people
-            # with possible unit related shenanigans, just dummy positions here.
-            f.write("geometry units bohr noautosym nocenter noautoz\n")
-            for i, atom in enumerate(atoms):
-                f.write(
-                    f"  {atom.symbol:<4s} {0.0:16.10f} {0.0:16.10f} {float(i):16.10f}\n"
-                )
-            f.write("end\n\n")
-
-            f.write(f"include {settings_path.name}\n\n")
-
-            f.write("driver\n")
-            if unix_mode:
-                f.write(f"  socket unix {socket_address}\n")
-            else:
[…]
