+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-rgpycrumbs unknown (2026-02-21)
-===============================
+rgpycrumbs v1.1.0 (2026-02-21)
+==============================
 
 Added
 -----

[…]
