+++ rgpycrumbs/orca/generate_orca_input.py
@@ -13,8 +13,8 @@ import sys
 from pathlib import Path
 
 import click
-from rich.logging import RichHandler
 from pychum import render_orca
+from rich.logging import RichHandler
 
 # --- Logging Configuration ---
 logging.basicConfig(
