### 2026-02-21 ad00e81
doc: regenerate README.md from readme_src.org [skip ci]
