+++ _sources/autoapi/rgpycrumbs/surfaces/index.rst.txt
@@ -4,15 +4,16 @@ rgpycrumbs.surfaces
 .. py:module:: rgpycrumbs.surfaces
 
 
-Attributes
+Submodules
 ----------
 
-.. autoapisummary::
+.. toctree::
+   :maxdepth: 1
 
-   rgpycrumbs.surfaces.k_matrix_matern_grad_map
-   rgpycrumbs.surfaces.k_matrix_se_grad_map
-   rgpycrumbs.surfaces.k_matrix_imq_grad_map
-   rgpycrumbs.surfaces.k_matrix_rq_grad_map
+   /autoapi/rgpycrumbs/surfaces/_base/index
+   /autoapi/rgpycrumbs/surfaces/_kernels/index
+   /autoapi/rgpycrumbs/surfaces/gradient/index
+   /autoapi/rgpycrumbs/surfaces/standard/index
 
 
 Classes
@@ -20,13 +21,16 @@ Classes
 
 .. autoapisummary::
 
-   rgpycrumbs.surfaces.FastTPS
-   rgpycrumbs.surfaces.FastMatern
-   rgpycrumbs.surfaces.GradientMatern
-   rgpycrumbs.surfaces.FastIMQ
-   rgpycrumbs.surfaces.GradientSE
+   rgpycrumbs.surfaces.BaseGradientSurface
+   rgpycrumbs.surfaces.BaseSurface
    rgpycrumbs.surfaces.GradientIMQ
+   rgpycrumbs.surfaces.GradientMatern
    rgpycrumbs.surfaces.GradientRQ
+   rgpycrumbs.surfaces.GradientSE
+   rgpycrumbs.surfaces.NystromGradientIMQ
+   rgpycrumbs.surfaces.FastIMQ
+   rgpycrumbs.surfaces.FastMatern
+   rgpycrumbs.surfaces.FastTPS
 
 
 Functions
@@ -34,149 +38,448 @@ Functions
 
 .. autoapisummary::
 
-   rgpycrumbs.surfaces.safe_cholesky_solve
    rgpycrumbs.surfaces.generic_negative_mll
-   rgpycrumbs.surfaces._tps_kernel_matrix
-   rgpycrumbs.surfaces.negative_mll_tps
-   rgpycrumbs.surfaces._tps_solve
-   rgpycrumbs.surfaces._tps_predict
-   rgpycrumbs.surfaces._matern_kernel_matrix
-   rgpycrumbs.surfaces.negative_mll_matern_std
-   rgpycrumbs.surfaces._matern_solve
-   rgpycrumbs.surfaces._matern_predict
-   rgpycrumbs.surfaces.matern_kernel_elem
-   rgpycrumbs.surfaces.full_covariance_matern
-   rgpycrumbs.surfaces.negative_mll_matern_grad
-   rgpycrumbs.surfaces._grad_matern_solve
-   rgpycrumbs.surfaces._grad_matern_predict
+   rgpycrumbs.surfaces.safe_cholesky_solve
    rgpycrumbs.surfaces._imq_kernel_matrix
-   rgpycrumbs.surfaces.negative_mll_imq_std
-   rgpycrumbs.surfaces._imq_solve
-   rgpycrumbs.surfaces._imq_predict
-   rgpycrumbs.surfaces.se_kernel_elem
-   rgpycrumbs.surfaces.full_covariance_se
-   rgpycrumbs.surfaces.negative_mll_se_grad
-   rgpycrumbs.surfaces._grad_se_solve
-   rgpycrumbs.surfaces._grad_se_predict
-   rgpycrumbs.surfaces.imq_kernel_elem
-   rgpycrumbs.surfaces.full_covariance_imq
[…]
