+++ _sources/autoapi/rgpycrumbs/eon/generate_nwchem_input/index.rst.txt
@@ -9,26 +9,11 @@ Functions
 
 .. autoapisummary::
 
-   rgpycrumbs.eon.generate_nwchem_input.generate_nwchem_input
    rgpycrumbs.eon.generate_nwchem_input.main
 
 
 Module Contents
 ---------------
-
-.. py:function:: generate_nwchem_input(pos_path: pathlib.Path, settings_path: pathlib.Path, socket_address: str, *, unix_mode: bool, mem_in_gb: int, output_path: pathlib.Path)
-
-   Generates a complete NWChem input file by combining atom types from a .con
-   file with a user-provided settings file.
-
-   Args:
-       pos_path: Path to the input geometry file (e.g., pos.con).
-       settings_path: Path to the user's NWChem settings file.
-       socket_address: The address for the socket (path for UNIX, host:port for TCP).
-       unix_mode: Boolean indicating if a UNIX socket should be used.
-       mem_in_gb: Memory to allocate for the NWChem calculation in Gigabytes.
-       output_path: The path for the generated NWChem input file.
-
 
 .. py:function:: main(pos_file: pathlib.Path, config: pathlib.Path, output: pathlib.Path)
 
