### 2026-02-22 e8c720a
feat(eon): add --source traj support to plt_neb
Add --source {eon,traj} and --input-traj options for plotting NEB profiles
and landscapes from extxyz trajectory files. Trajectory data is converted
to the same format as eOn .dat/.con pairs via chemparseplot.parse.chemgp,
reusing all existing plot_energy_path and plot_landscape_surface code.
