+++ rgpycrumbs/eon/plt_neb.py
@@ -51,6 +51,11 @@ import matplotlib.pyplot as plt
 import numpy as np
 import polars as pl
 from adjustText import adjust_text
+from chemparseplot.parse.chemgp.neb import (
+    load_trajectory,
+    trajectory_to_landscape_df,
+    trajectory_to_profile_dat,
+)
 from chemparseplot.parse.eon.neb import (
     aggregate_neb_landscape_data,
     compute_profile_rmsd,
@@ -136,6 +141,18 @@ IRA_KMAX_DEFAULT = 1.8
     type=click.Path(exists=False, dir_okay=False, path_type=Path),
     default=Path("sp.con"),
     help="Path to explicit saddle point file (eOn sp.con).",
+)
+@click.option(
+    "--source",
+    type=click.Choice(["eon", "traj"]),
+    default="eon",
+    help="Data source: 'eon' for .dat/.con pairs, 'traj' for extxyz trajectory.",
+)
+@click.option(
+    "--input-traj",
+    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+    default=None,
+    help="Path to extxyz trajectory file (used with --source traj).",
 )
 @click.option(
     "--plot-type",
@@ -366,6 +383,9 @@ def main(
     input_path_pattern,
     con_file,
     additional_con,
+    # --- Data Source ---
+    source,
+    input_traj,
     # --- Plot Behavior ---
     plot_type,
     landscape_mode,
@@ -480,38 +500,53 @@ def main(
                 log.critical("Cannot proceed without structures. Exiting.")
                 sys.exit(1)
 
+    # --- Trajectory source: load once if applicable ---
+    traj_atoms_list = None
+    if source == "traj":
+        if not input_traj:
+            log.critical("--input-traj is required when --source traj is used.")
+            sys.exit(1)
+        traj_atoms_list = load_trajectory(str(input_traj))
+
     if plot_type == "landscape":
         # --- Landscape Plot ---
-        dat_paths = find_file_paths(input_dat_pattern)
-        con_paths = find_file_paths(str(input_path_pattern))
-
-        if not dat_paths:
-            log.critical(f"No data files found for pattern: {input_dat_pattern}")
-            sys.exit(1)
-
-        # Fallback if no path files found but main file exists
-        if not con_paths and con_file:
-            con_paths = [con_file]
-
-        y_col = 2 if plot_mode == "energy" else 4
         z_label = (
             "Relative Energy (eV)"
             if plot_mode == "energy"
             else r"Lowest Eigenvalue (eV/$\AA^2$)"
         )
 
-        df = aggregate_neb_landscape_data(
-            dat_paths,
-            con_paths,
-            y_col,
-            None,
-            cache_file,
-            force_recompute,
-            ira_kmax,
-            augment_dat=augment_dat,
-            augment_con=augment_con,
-            ref_atoms=atoms_list[0] if atoms_list else None,  # main reactant
-            prod_atoms=atoms_list[-1] if atoms_list else None,  # main product
+        if source == "traj":
+            df = trajectory_to_landscape_df(traj_atoms_list, ira_kmax=ira_kmax)
+            # Use traj structures for con_file features when not provided
+            if atoms_list is None:
+                atoms_list = traj_atoms_list
+        else:
+            dat_paths = find_file_paths(input_dat_pattern)
+            con_paths = find_file_paths(str(input_path_pattern))
+
+            if not dat_paths:
+                log.critical(f"No data files found for pattern: {input_dat_pattern}")
+                sys.exit(1)
+
+            # Fallback if no path files found but main file exists
+            if not con_paths and con_file:
+                con_paths = [con_file]
+
+            y_col = 2 if plot_mode == "energy" else 4
+
[…]
