### 2026-02-23 52aa4cc
refactor(eon): update chemparseplot import path for trajectory rename
The chemparseplot parse.chemgp module was renamed to parse.trajectory
to reflect that the parser is generic (any ASE-readable trajectory),
not ChemGP-specific.
