### 2026-02-23 493d53d
feat(eon): add --source hdf5 support to plt_neb
Add ChemGP HDF5 as a data source for NEB plotting. Reads neb_result.h5
or neb_history.h5 via chemparseplot.parse.trajectory.hdf5 for both
profile and landscape plot types, with history-to-result fallback.
