### 2026-02-23 acb13e1
doc: add versionadded directives to all public API docstrings
Annotate every public function, class, and module with RST
versionadded directives so autoapi-generated API docs show
introduction versions.
