### 2026-02-23 b1fa50c
feat(eon): add --n-inducing flag and fix (s,d) axis labels
Adds --n-inducing CLI option for Nystrom inducing point count,
renames axis labels to Reaction progress s / Orthogonal deviation d,
fixes legend placement (loc=best) for projected mode, and reorders
tight_layout before strip repositioning to prevent xlabel overlap.
