+++ rgpycrumbs/eon/plt_neb.py
@@ -52,17 +52,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 import polars as pl
 from adjustText import adjust_text
-from chemparseplot.parse.trajectory.hdf5 import (
-    history_to_landscape_df as hdf5_history_to_landscape_df,
-    history_to_profile_dats,
-    result_to_atoms_list,
-    result_to_profile_dat as hdf5_result_to_profile_dat,
-)
-from chemparseplot.parse.trajectory.neb import (
-    load_trajectory,
-    trajectory_to_landscape_df,
-    trajectory_to_profile_dat,
-)
 from chemparseplot.parse.eon.neb import (
     aggregate_neb_landscape_data,
     compute_profile_rmsd,
@@ -72,6 +61,21 @@ from chemparseplot.parse.eon.neb import (
 
 # --- Library Imports ---
 from chemparseplot.parse.file_ import find_file_paths
+from chemparseplot.parse.trajectory.hdf5 import (
+    history_to_landscape_df as hdf5_history_to_landscape_df,
+)
+from chemparseplot.parse.trajectory.hdf5 import (
+    history_to_profile_dats,
+    result_to_atoms_list,
+)
+from chemparseplot.parse.trajectory.hdf5 import (
+    result_to_profile_dat as hdf5_result_to_profile_dat,
+)
+from chemparseplot.parse.trajectory.neb import (
+    load_trajectory,
+    trajectory_to_landscape_df,
+    trajectory_to_profile_dat,
+)
 from chemparseplot.plot.neb import (
     plot_energy_path,
     plot_landscape_path_overlay,
@@ -227,6 +231,12 @@ IRA_KMAX_DEFAULT = 1.8
     ),
     default="rbf",
     help="Interpolation method for the 2D surface.",
+)
+@click.option(
+    "--n-inducing",
+    type=int,
+    default=None,
+    help="Number of inducing points for Nystrom or RFF features. Defaults to 300 (Nystrom) or 500 (RFF).",
 )
 @click.option(
     "--show-pts/--no-show-pts",
@@ -411,6 +421,7 @@ def main(
     show_pts,
     plot_mode,
     surface_type,
+    n_inducing,
     # --- Output & Slicing ---
     output_file,
     start,
@@ -551,6 +562,7 @@ def main(
                 from chemparseplot.parse.trajectory.neb import (
                     trajectory_to_landscape_df as _traj_ldf,
                 )
+
                 hdf5_atoms = _r2a(h5_str)
                 df = _traj_ldf(hdf5_atoms, ira_kmax=ira_kmax)
             if atoms_list is None:
@@ -581,7 +593,7 @@ def main(
                 augment_con=augment_con,
                 ref_atoms=atoms_list[0] if atoms_list else None,  # main reactant
                 prod_atoms=atoms_list[-1] if atoms_list else None,  # main product
-        )
+            )
 
         # Surface Generation
         if landscape_mode == "surface":
@@ -627,6 +639,7 @@ def main(
                 variance_threshold=0.5,  # 50% uncertainty
                 project_path=project_path,
                 extra_points=extra_pts_arr,
+                n_inducing=n_inducing,
             )
 
         # Path Overlay (Final Step)
@@ -811,8 +824,8 @@ def main(
 
         # Labels
         if project_path:
-            final_xlabel = xlabel or r"RMSD from Reactant ($\AA$)"
-            final_ylabel = ylabel or r"Orthogonal Deviation $d$ ($\AA$)"
+            final_xlabel = xlabel or r"Reaction progress $s$ ($\AA$)"
+            final_ylabel = ylabel or r"Orthogonal deviation $d$ ($\AA$)"
[…]
