### 2026-02-23 f8876b7
feat(surfaces): export Nystrom constants, alpha-differentiated scatter
Add NYSTROM_THRESHOLD, NYSTROM_N_INDUCING_DEFAULT, and
nystrom_paths_needed() to rgpycrumbs.surfaces so downstream
consumers share a single source of truth for the threshold
and inducing budget.
