### 2026-02-23 1427ef8
perf(geom): parallelize per-image RMSD alignment over threads
Each image alignment (IRA + Procrustes fallback) is independent
and the Fortran/numpy calls release the GIL, so ThreadPoolExecutor
scales linearly with available cores.
