### 2026-02-23 e12b660
test: add parallel RMSD and strip-renderer CLI tests
TestParallelRMSD covers reference zero, non-negativity, monotonicity,
parallel vs sequential equivalence, single image edge case, product
reference, and IRA variant. Strip renderer test verifies --strip-renderer
appears in plt-neb help output.
