### 2026-02-24 953dfb3
doc: add newsfragments for v1.2.0 features
