+++ docs/newsfragments/7.dev.rst
@@ -1,0 +1,1 @@
+Added versionadded directives to all public API docstrings.
