+++ docs/newsfragments/1.added.rst
@@ -1,0 +1,1 @@
+Added --strip-renderer CLI option for selecting structure rendering backend.
