+++ docs/newsfragments/5.added.rst
@@ -1,0 +1,1 @@
+Added --n-inducing flag to eON surface-fitting CLI for controlling the Nystrom approximation inducing-point count.
