### 2026-02-24 b459f84
rgpycrumbs: prepare v1.2.0 release
