+++ docs/newsfragments/4.added.rst
@@ -1,1 +1,0 @@
-Added --source hdf5 support to plt_neb for loading NEB paths from ChemGP HDF5 output files.
