+++ docs/newsfragments/3.added.rst
@@ -1,1 +1,0 @@
-Added --source traj support to plt_neb for loading NEB paths from ASE-readable trajectory files (extxyz, .traj, etc.).
