+++ docs/newsfragments/6.changed.rst
@@ -1,1 +1,0 @@
-Per-image RMSD alignment in rgpycrumbs.geom now runs in parallel over threads for improved performance on large NEB paths.
