### 2026-02-24 59ff3a8
fix(test): move plt_neb CLI test to eon marker, fix ruff noqa
The strip-renderer test imported plt_neb (heavy deps) under the pure
marker, causing ModuleNotFoundError for polars in the uv test matrix.
Moved the test to tests/test_eon_cli.py with proper skip_if_not_env
guard and eon marker. Added PLC0415 noqa for intentional late imports.
