### 2026-02-24 4bf8518
doc: regenerate README.md from readme_src.org [skip ci]
