+++ README.md
@@ -2,7 +2,7 @@
 # Table of Contents
 
 -   [About](#about)
-    -   [Ecosystem Overview](#org723df51)
+    -   [Ecosystem Overview](#org7c2c5ed)
     -   [CLI Design Philosophy](#cli-how)
 -   [Usage](#usage)
     -   [Library API](#library-api)
@@ -10,8 +10,8 @@
         -   [eOn](#cli-eon)
 -   [Contributing](#contributing)
     -   [Development](#development)
-        -   [When is pixi needed?](#org1e75565)
-        -   [Versioning](#org4873ddd)
+        -   [When is pixi needed?](#org4f007ea)
+        -   [Versioning](#org3ff67ac)
     -   [Release Process](#release-notes)
 -   [License](#license)
 
@@ -24,6 +24,7 @@
 ![img](https://raw.githubusercontent.com/HaoZeke/rgpycrumbs/refs/heads/main/branding/logo/pycrumbs_logo.webp)
 
 [![Hatch project](https://img.shields.io/badge/%F0%9F%A5%9A-Hatch-4051b5.svg)](https://github.com/pypa/hatch)
+[![DOI](https://zenodo.org/badge/795596895.svg)](https://doi.org/10.5281/zenodo.18529798)
 
 A **pure-python** computational library and CLI toolkit for chemical physics
 research. `rgpycrumbs` provides both importable library modules for
@@ -41,7 +42,7 @@ The library side offers:
 The CLI tools rely on optional dependencies fetched on-demand via PEP 723 + `uv`.
 
 
-<a id="org723df51"></a>
+<a id="org7c2c5ed"></a>
 
 ## Ecosystem Overview
 
@@ -97,13 +98,13 @@ The library modules can be imported directly:
     # Surface fitting (requires jax: pip install rgpycrumbs[surfaces])
     from rgpycrumbs.surfaces import get_surface_model
     model = get_surface_model("tps")
-
+    
     # Structure analysis (requires ase, scipy: pip install rgpycrumbs[analysis])
     from rgpycrumbs.geom.analysis import analyze_structure
-
+    
     # Spline interpolation (requires scipy: pip install rgpycrumbs[interpolation])
     from rgpycrumbs.interpolation import spline_interp
-
+    
     # Data types (no extra deps)
     from rgpycrumbs.basetypes import nebpath, SaddleMeasure
 
@@ -120,12 +121,12 @@ You can see the list of available command groups:
 
     $ python -m rgpycrumbs.cli --help
     Usage: rgpycrumbs [OPTIONS] COMMAND [ARGS]...
-
+    
       A dispatcher that runs self-contained scripts using 'uv'.
-
+    
     Options:
       --help  Show this message and exit.
-
+    
     Commands:
       eon  Dispatches to a script within the 'eon' submodule.
 
@@ -137,13 +138,13 @@ You can see the list of available command groups:
 -   Plotting NEB Paths (`plt-neb`)
 
     This script visualizes the energy profile of Nudged Elastic Band (NEB) calculations over optimization steps.
-
+    
     To see the help text for this specific script:
-
+    
         $ python -m rgpycrumbs eon plt-neb --help
         --> Dispatching to: uv run /path/to/rgpycrumbs/eon/plt_neb.py --help
         Usage: plt_neb.py [OPTIONS]
-
+        
           Plots a series of NEB energy paths from .dat files.
         ...
         Options:
@@ -152,13 +153,13 @@ You can see the list of available command groups:
           --start INTEGER           Starting file index to plot (inclusive).
           --end INTEGER             Ending file index to plot (exclusive).
           --help                    Show this message and exit.
-
+    
[…]
