+++ _sources/autoapi/rgpycrumbs/plumed/direct_reconstruction/index.rst.txt
@@ -45,6 +45,8 @@ Module Contents
 
    Calculates a 1D or 2D Free Energy Surface (FES) from a metadynamics hills file.
 
+   .. versionadded:: 0.0.3
+
    This is a Python/NumPy translation of the R function fes2.hillsfile.
 
    Args:
@@ -83,6 +85,8 @@ Module Contents
 .. py:function:: find_fes_minima(fes_result, nbins=8)
 
    Finds local minima on a Free Energy Surface by dividing it into bins.
+
+   .. versionadded:: 0.0.3
 
    This is a Python/Pandas translation of the R function fesminima.fes.
 
