+++ _sources/autoapi/rgpycrumbs/geom/analysis/index.rst.txt
@@ -20,6 +20,8 @@ Module Contents
    Analyzes an ASE Atoms object to calculate the distance matrix, bond matrix,
    identify molecular fragments, and calculate inter-fragment distances.
 
+   .. versionadded:: 1.0.0
+
    Args:
        atoms: The ASE Atoms object to analyze.
        covalent_scale:  A scaling factor applied to the covalent radii to
