+++ _sources/autoapi/helpers/index.rst.txt
@@ -17,3 +17,8 @@ Module Contents
 
 .. py:function:: one_day_tdelta(etime: str, stime: str, format_time: str = '%H:%M:%S')
 
+   Compute the timedelta between two times within the same day.
+
+   .. versionadded:: 0.0.2
+
+
