### 2026-02-24 e963e12
chore(lint): sort __all__ in surfaces
